Adding UI window formatting to 	renamed:    src/Python PYSAT_GUI/PythonUI/mainframe_ok.py -> src/Python PYSAT_GUI/PythonUI/mainframe_.py 	new file:   src/Python PYSAT_GUI/PythonUI/regression_.py 	new file:   src/PYSAT_Gui_UI_Forms/01_mainwindow_get_data.ui 	new file:   src/PYSAT_Gui_UI_Forms/PythonUI/01_mainwindow_get_data.py 	new file:   src/Python PYSAT_GUI/PythonUI/file_out_.py 	new file:   src/Python PYSAT_GUI/PythonUI/get_data_.py Fixing some UI files 	modified:   src/PYSAT_Gui_UI_Forms/01_mainwindow_file_out.ui 	modified:   src/PYSAT_Gui_UI_Forms/PythonUI/01_mainwindow_file_out.py Scratch file, is temporarily being used to help me organize the new code 	modified:   src/Python PYSAT_GUI/.recycle.bin/py_scratch.py Commenting various parts of Main for debugging and testing purposes, It will be fixed once everything runs correctly 	modified:   src/Python PYSAT_GUI/pysat_main.py 	modified:   src/Python PYSAT_GUI/pysat_ui.py This became a useless file, it wasn't adding anything new 	deleted:    src/PYSAT_Gui_UI_Forms/PythonUI/total.py These technically should be renamed files, but perhaps to many things changed that prevented it from being called <renamed> 	deleted:    src/Python PYSAT_GUI/PythonUI/mainwindow_file_out.py 	deleted:    src/Python PYSAT_GUI/PythonUI/mainwindow_regression.py
