+++ src/Python PYSAT_GUI/.recycle.bin/py_scratch.py
@@ -1,594 +1,5 @@
 
 
-    def normalization(self, MainWindow):
-        self.Normalization = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.Normalization.setFont(font)
-        self.Normalization.setObjectName(_fromUtf8("Normalization"))
-        self.verticalLayout_27 = QtGui.QVBoxLayout(self.Normalization)
-        self.verticalLayout_27.setMargin(11)
-        self.verticalLayout_27.setSpacing(6)
-        self.verticalLayout_27.setObjectName(_fromUtf8("verticalLayout_27"))
-        self.verticalLayout_26 = QtGui.QVBoxLayout()
-        self.verticalLayout_26.setMargin(11)
-        self.verticalLayout_26.setSpacing(6)
-        self.verticalLayout_26.setObjectName(_fromUtf8("verticalLayout_26"))
-        self.verticalLayout_2 = QtGui.QVBoxLayout()
-        self.verticalLayout_2.setMargin(11)
-        self.verticalLayout_2.setSpacing(6)
-        self.verticalLayout_2.setObjectName(_fromUtf8("verticalLayout_2"))
-        self.horizontalLayout_2 = QtGui.QHBoxLayout()
-        self.horizontalLayout_2.setMargin(11)
-        self.horizontalLayout_2.setSpacing(6)
-        self.horizontalLayout_2.setObjectName(_fromUtf8("horizontalLayout_2"))
-        self.norm_label = QtGui.QLabel(self.Normalization)
-        self.norm_label.setObjectName(_fromUtf8("norm_label"))
-        self.horizontalLayout_2.addWidget(self.norm_label)
-        self.norm_spinBox_9 = QtGui.QSpinBox(self.Normalization)
-        self.norm_spinBox_9.setMaximum(1000)
-        self.norm_spinBox_9.setObjectName(_fromUtf8("norm_spinBox_9"))
-        self.horizontalLayout_2.addWidget(self.norm_spinBox_9)
-        self.norm_spinBox = QtGui.QSpinBox(self.Normalization)
-        self.norm_spinBox.setMaximum(1000)
-        self.norm_spinBox.setObjectName(_fromUtf8("norm_spinBox"))
-        self.horizontalLayout_2.addWidget(self.norm_spinBox)
-        self.verticalLayout_2.addLayout(self.horizontalLayout_2)
-        self.verticalLayout_26.addLayout(self.verticalLayout_2)
-        self.verticalLayout_27.addLayout(self.verticalLayout_26)
-        self.horizontalLayout_25 = QtGui.QHBoxLayout()
-        self.horizontalLayout_25.setMargin(11)
-        self.horizontalLayout_25.setSpacing(6)
-        self.horizontalLayout_25.setObjectName(_fromUtf8("horizontalLayout_25"))
-        spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.horizontalLayout_25.addItem(spacerItem)
-        self.NormValuebutton = QtGui.QPushButton(self.Normalization)
-        self.NormValuebutton.setObjectName(_fromUtf8("NormValuebutton"))
-        self.horizontalLayout_25.addWidget(self.NormValuebutton)
-        self.verticalLayout_27.addLayout(self.horizontalLayout_25)
-        self.verticalLayout_8.addWidget(self.Normalization)
-        self.Normalization.raise_()
-        self.Normalization.setTitle(_translate("MainWindow", "Normalization", None))
-        self.norm_label.setText(_translate("MainWindow", "Value 1", None))
-        self.NormValuebutton.setText(_translate("MainWindow", "Add Value", None))
-        self.normalization_button = QtGui.QPushButton(self.Normalization)
-        self.set_module_button(self.verticalLayout_27, self.normalization_button)
-        try:
-            self.normalization_button.clicked.connect(
-                lambda: self.pysat_fun.get_ranges(([(int(self.norm_spinBox_9.text()), int(self.norm_spinBox.text()))])))
-        except:
-            pass
-
-    def comp_ranges(self, MainWindow):
-        self.CompRanges = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.CompRanges.setFont(font)
-        self.CompRanges.setObjectName(_fromUtf8("CompRanges"))
-        self.verticalLayout_7 = QtGui.QVBoxLayout(self.CompRanges)
-        self.verticalLayout_7.setMargin(11)
-        self.verticalLayout_7.setSpacing(6)
-        self.verticalLayout_7.setObjectName(_fromUtf8("verticalLayout_7"))
-        self.verticalLayout_4 = QtGui.QVBoxLayout()
-        self.verticalLayout_4.setMargin(11)
-        self.verticalLayout_4.setSpacing(6)
-        self.verticalLayout_4.setObjectName(_fromUtf8("verticalLayout_4"))
-        self.horizontalLayout_3 = QtGui.QHBoxLayout()
-        self.horizontalLayout_3.setMargin(11)
-        self.horizontalLayout_3.setSpacing(6)
-        self.horizontalLayout_3.setObjectName(_fromUtf8("horizontalLayout_3"))
-        self.label_80 = QtGui.QLabel(self.CompRanges)
-        self.label_80.setObjectName(_fromUtf8("label_80"))
-        self.horizontalLayout_3.addWidget(self.label_80)
-        self.comp_range = QtGui.QSpinBox(self.CompRanges)
-        self.comp_range.setMinimum(-100)
-        self.comp_range.setMaximum(1000)
-        self.comp_range.setObjectName(_fromUtf8("comp_range"))
-        self.horizontalLayout_3.addWidget(self.comp_range)
-        self.comp_range_2 = QtGui.QSpinBox(self.CompRanges)
-        self.comp_range_2.setMaximum(1000)
-        self.comp_range_2.setObjectName(_fromUtf8("comp_range_2"))
-        self.horizontalLayout_3.addWidget(self.comp_range_2)
-        self.verticalLayout_4.addLayout(self.horizontalLayout_3)
-        self.horizontalLayout_5 = QtGui.QHBoxLayout()
-        self.horizontalLayout_5.setMargin(11)
-        self.horizontalLayout_5.setSpacing(6)
-        self.horizontalLayout_5.setObjectName(_fromUtf8("horizontalLayout_5"))
-        self.label_81 = QtGui.QLabel(self.CompRanges)
-        self.label_81.setObjectName(_fromUtf8("label_81"))
-        self.horizontalLayout_5.addWidget(self.label_81)
-        self.comp_range_3 = QtGui.QSpinBox(self.CompRanges)
-        self.comp_range_3.setMaximum(1000)
-        self.comp_range_3.setObjectName(_fromUtf8("comp_range_3"))
-        self.horizontalLayout_5.addWidget(self.comp_range_3)
-        self.comp_range_4 = QtGui.QSpinBox(self.CompRanges)
-        self.comp_range_4.setMaximum(1000)
-        self.comp_range_4.setObjectName(_fromUtf8("comp_range_4"))
-        self.horizontalLayout_5.addWidget(self.comp_range_4)
-        self.verticalLayout_4.addLayout(self.horizontalLayout_5)
-        self.horizontalLayout_7 = QtGui.QHBoxLayout()
-        self.horizontalLayout_7.setMargin(11)
-        self.horizontalLayout_7.setSpacing(6)
-        self.horizontalLayout_7.setObjectName(_fromUtf8("horizontalLayout_7"))
-        self.label_82 = QtGui.QLabel(self.CompRanges)
-        self.label_82.setObjectName(_fromUtf8("label_82"))
-        self.horizontalLayout_7.addWidget(self.label_82)
-        self.comp_range_5 = QtGui.QSpinBox(self.CompRanges)
-        self.comp_range_5.setMaximum(1000)
-        self.comp_range_5.setObjectName(_fromUtf8("comp_range_5"))
-        self.horizontalLayout_7.addWidget(self.comp_range_5)
-        self.comp_range_6 = QtGui.QSpinBox(self.CompRanges)
-        self.comp_range_6.setMaximum(1000)
-        self.comp_range_6.setObjectName(_fromUtf8("comp_range_6"))
-        self.horizontalLayout_7.addWidget(self.comp_range_6)
-        self.verticalLayout_4.addLayout(self.horizontalLayout_7)
-        self.horizontalLayout_10 = QtGui.QHBoxLayout()
-        self.horizontalLayout_10.setMargin(11)
-        self.horizontalLayout_10.setSpacing(6)
-        self.horizontalLayout_10.setObjectName(_fromUtf8("horizontalLayout_10"))
-        self.label_85 = QtGui.QLabel(self.CompRanges)
-        self.label_85.setObjectName(_fromUtf8("label_85"))
-        self.horizontalLayout_10.addWidget(self.label_85)
-        self.comp_range_7 = QtGui.QSpinBox(self.CompRanges)
-        self.comp_range_7.setMaximum(1000)
-        self.comp_range_7.setObjectName(_fromUtf8("comp_range_7"))
-        self.horizontalLayout_10.addWidget(self.comp_range_7)
-        self.comp_range_8 = QtGui.QSpinBox(self.CompRanges)
-        self.comp_range_8.setMaximum(1000)
-        self.comp_range_8.setObjectName(_fromUtf8("comp_range_8"))
-        self.horizontalLayout_10.addWidget(self.comp_range_8)
-        self.verticalLayout_4.addLayout(self.horizontalLayout_10)
-        self.verticalLayout_7.addLayout(self.verticalLayout_4)
-        self.horizontalLayout_9 = QtGui.QHBoxLayout()
-        self.horizontalLayout_9.setMargin(11)
-        self.horizontalLayout_9.setSpacing(6)
-        self.horizontalLayout_9.setObjectName(_fromUtf8("horizontalLayout_9"))
-        self.label_40 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_40.setFont(font)
-        self.label_40.setObjectName(_fromUtf8("label_40"))
-        self.horizontalLayout_9.addWidget(self.label_40)
-        self.horizontalLayout_6 = QtGui.QHBoxLayout()
-        self.horizontalLayout_6.setMargin(11)
-        self.horizontalLayout_6.setSpacing(6)
-        self.horizontalLayout_6.setObjectName(_fromUtf8("horizontalLayout_6"))
-        self.label_20 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_20.setFont(font)
-        self.label_20.setObjectName(_fromUtf8("label_20"))
-        self.horizontalLayout_6.addWidget(self.label_20)
-        self.label_21 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_21.setFont(font)
-        self.label_21.setObjectName(_fromUtf8("label_21"))
-        self.horizontalLayout_6.addWidget(self.label_21)
-        self.label_22 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_22.setFont(font)
-        self.label_22.setObjectName(_fromUtf8("label_22"))
-        self.horizontalLayout_6.addWidget(self.label_22)
-        self.label_23 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_23.setFont(font)
-        self.label_23.setObjectName(_fromUtf8("label_23"))
-        self.horizontalLayout_6.addWidget(self.label_23)
-        self.label_24 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_24.setFont(font)
-        self.label_24.setObjectName(_fromUtf8("label_24"))
-        self.horizontalLayout_6.addWidget(self.label_24)
-        self.label_27 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_27.setFont(font)
-        self.label_27.setObjectName(_fromUtf8("label_27"))
-        self.horizontalLayout_6.addWidget(self.label_27)
-        self.label_29 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_29.setFont(font)
-        self.label_29.setObjectName(_fromUtf8("label_29"))
-        self.horizontalLayout_6.addWidget(self.label_29)
-        self.label_28 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_28.setFont(font)
-        self.label_28.setObjectName(_fromUtf8("label_28"))
-        self.horizontalLayout_6.addWidget(self.label_28)
-        self.label_26 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_26.setFont(font)
-        self.label_26.setObjectName(_fromUtf8("label_26"))
-        self.horizontalLayout_6.addWidget(self.label_26)
-        self.label_25 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_25.setFont(font)
-        self.label_25.setObjectName(_fromUtf8("label_25"))
-        self.horizontalLayout_6.addWidget(self.label_25)
-        self.label_32 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_32.setFont(font)
-        self.label_32.setObjectName(_fromUtf8("label_32"))
-        self.horizontalLayout_6.addWidget(self.label_32)
-        self.label_30 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_30.setFont(font)
-        self.label_30.setObjectName(_fromUtf8("label_30"))
-        self.horizontalLayout_6.addWidget(self.label_30)
-        self.label_33 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_33.setFont(font)
-        self.label_33.setObjectName(_fromUtf8("label_33"))
-        self.horizontalLayout_6.addWidget(self.label_33)
-        self.label_34 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_34.setFont(font)
-        self.label_34.setObjectName(_fromUtf8("label_34"))
-        self.horizontalLayout_6.addWidget(self.label_34)
-        self.label_31 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_31.setFont(font)
-        self.label_31.setObjectName(_fromUtf8("label_31"))
-        self.horizontalLayout_6.addWidget(self.label_31)
-        self.label_39 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_39.setFont(font)
-        self.label_39.setObjectName(_fromUtf8("label_39"))
-        self.horizontalLayout_6.addWidget(self.label_39)
-        self.label_36 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_36.setFont(font)
-        self.label_36.setObjectName(_fromUtf8("label_36"))
-        self.horizontalLayout_6.addWidget(self.label_36)
-        self.label_37 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_37.setFont(font)
-        self.label_37.setObjectName(_fromUtf8("label_37"))
-        self.horizontalLayout_6.addWidget(self.label_37)
-        self.label_35 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_35.setFont(font)
-        self.label_35.setObjectName(_fromUtf8("label_35"))
-        self.horizontalLayout_6.addWidget(self.label_35)
-        self.label_38 = QtGui.QLabel(self.CompRanges)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.label_38.setFont(font)
-        self.label_38.setObjectName(_fromUtf8("label_38"))
-        self.horizontalLayout_6.addWidget(self.label_38)
-        self.horizontalLayout_9.addLayout(self.horizontalLayout_6)
-        self.verticalLayout_7.addLayout(self.horizontalLayout_9)
-        self.verticalLayout_8.addWidget(self.CompRanges)
-        self.CompRanges.raise_()
-        QtCore.QObject.connect(self.comp_range, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")), self.label_21.setNum)
-        QtCore.QObject.connect(self.comp_range_2, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")), self.label_23.setNum)
-        QtCore.QObject.connect(self.comp_range_3, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")), self.label_29.setNum)
-        QtCore.QObject.connect(self.comp_range_4, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")), self.label_26.setNum)
-        QtCore.QObject.connect(self.comp_range_5, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")), self.label_30.setNum)
-        QtCore.QObject.connect(self.comp_range_6, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")), self.label_34.setNum)
-        QtCore.QObject.connect(self.comp_range_7, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")), self.label_36.setNum)
-        QtCore.QObject.connect(self.comp_range_8, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")), self.label_35.setNum)
-        self.CompRanges.setTitle(_translate("MainWindow", "Submodel Ranges", None))
-        self.label_80.setText(_translate("MainWindow", "Value 1", None))
-        self.label_81.setText(_translate("MainWindow", "Value 2", None))
-        self.label_82.setText(_translate("MainWindow", "Value 3", None))
-        self.label_85.setText(_translate("MainWindow", "Full Model Value", None))
-        self.label_40.setText(_translate("MainWindow", "compranges", None))
-        self.label_20.setText(_translate("MainWindow", "[", None))
-        self.label_21.setText(_translate("MainWindow", "0", None))
-        self.label_22.setText(_translate("MainWindow", ",", None))
-        self.label_23.setText(_translate("MainWindow", "0", None))
-        self.label_24.setText(_translate("MainWindow", "]", None))
-        self.label_27.setText(_translate("MainWindow", "[", None))
-        self.label_29.setText(_translate("MainWindow", "0", None))
-        self.label_28.setText(_translate("MainWindow", ",", None))
-        self.label_26.setText(_translate("MainWindow", "0", None))
-        self.label_25.setText(_translate("MainWindow", "]", None))
-        self.label_32.setText(_translate("MainWindow", "[", None))
-        self.label_30.setText(_translate("MainWindow", "0", None))
-        self.label_33.setText(_translate("MainWindow", ",", None))
-        self.label_34.setText(_translate("MainWindow", "0", None))
-        self.label_31.setText(_translate("MainWindow", "]", None))
-        self.label_39.setText(_translate("MainWindow", "[", None))
-        self.label_36.setText(_translate("MainWindow", "0", None))
-        self.label_37.setText(_translate("MainWindow", ",", None))
-        self.label_35.setText(_translate("MainWindow", "0", None))
-        self.label_38.setText(_translate("MainWindow", "]", None))
-        self.compRanges_button = QtGui.QPushButton(self.CompRanges)
-        self.set_module_button(self.verticalLayout_7, self.compRanges_button)
-        try:
-            self.compRanges_button.clicked.connect(
-                lambda: self.pysat_fun.set_compranges([[int(self.comp_range.text()), int(self.comp_range_2.text())],
-                                                       [int(self.comp_range_3.text()), int(self.comp_range_4.text())],
-                                                       [int(self.comp_range_5.text()), int(self.comp_range_6.text())],
-                                                       [int(self.comp_range_7.text()), int(self.comp_range_8.text())]]))
-        except:
-            pass
-
-    def num_of_components(self, MainWindow):
-        self.NumberComponents = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.NumberComponents.setFont(font)
-        self.NumberComponents.setObjectName(_fromUtf8("NumberComponents"))
-        self.verticalLayout_17 = QtGui.QVBoxLayout(self.NumberComponents)
-        self.verticalLayout_17.setMargin(11)
-        self.verticalLayout_17.setSpacing(6)
-        self.verticalLayout_17.setObjectName(_fromUtf8("verticalLayout_17"))
-        self.horizontalLayout_14 = QtGui.QHBoxLayout()
-        self.horizontalLayout_14.setMargin(11)
-        self.horizontalLayout_14.setSpacing(6)
-        self.horizontalLayout_14.setObjectName(_fromUtf8("horizontalLayout_14"))
-        self.verticalLayout_18 = QtGui.QVBoxLayout()
-        self.verticalLayout_18.setMargin(11)
-        self.verticalLayout_18.setSpacing(6)
-        self.verticalLayout_18.setObjectName(_fromUtf8("verticalLayout_18"))
-        self.label_70 = QtGui.QLabel(self.NumberComponents)
-        self.label_70.setObjectName(_fromUtf8("label_70"))
-        self.verticalLayout_18.addWidget(self.label_70)
-        self.label_71 = QtGui.QLabel(self.NumberComponents)
-        self.label_71.setObjectName(_fromUtf8("label_71"))
-        self.verticalLayout_18.addWidget(self.label_71)
-        self.label_72 = QtGui.QLabel(self.NumberComponents)
-        self.label_72.setObjectName(_fromUtf8("label_72"))
-        self.verticalLayout_18.addWidget(self.label_72)
-        self.label_73 = QtGui.QLabel(self.NumberComponents)
-        self.label_73.setObjectName(_fromUtf8("label_73"))
-        self.verticalLayout_18.addWidget(self.label_73)
-        self.horizontalLayout_14.addLayout(self.verticalLayout_18)
-        self.verticalLayout_19 = QtGui.QVBoxLayout()
-        self.verticalLayout_19.setMargin(11)
-        self.verticalLayout_19.setSpacing(6)
-        self.verticalLayout_19.setObjectName(_fromUtf8("verticalLayout_19"))
-        self.create_model_spin = QtGui.QSpinBox(self.NumberComponents)
-        self.create_model_spin.setMaximum(20)
-        self.create_model_spin.setObjectName(_fromUtf8("create_model_spin"))
-        self.verticalLayout_19.addWidget(self.create_model_spin)
-        self.create_model_spin_2 = QtGui.QSpinBox(self.NumberComponents)
-        self.create_model_spin_2.setMaximum(20)
-        self.create_model_spin_2.setObjectName(_fromUtf8("create_model_spin_2"))
-        self.verticalLayout_19.addWidget(self.create_model_spin_2)
-        self.create_model_spin_3 = QtGui.QSpinBox(self.NumberComponents)
-        self.create_model_spin_3.setMaximum(20)
-        self.create_model_spin_3.setObjectName(_fromUtf8("create_model_spin_3"))
-        self.verticalLayout_19.addWidget(self.create_model_spin_3)
-        self.create_model_spin_4 = QtGui.QSpinBox(self.NumberComponents)
-        self.create_model_spin_4.setMaximum(20)
-        self.create_model_spin_4.setObjectName(_fromUtf8("create_model_spin_4"))
-        self.verticalLayout_19.addWidget(self.create_model_spin_4)
-        self.horizontalLayout_14.addLayout(self.verticalLayout_19)
-        self.verticalLayout_17.addLayout(self.horizontalLayout_14)
-        self.horizontalLayout_15 = QtGui.QHBoxLayout()
-        self.horizontalLayout_15.setMargin(11)
-        self.horizontalLayout_15.setSpacing(6)
-        self.horizontalLayout_15.setObjectName(_fromUtf8("horizontalLayout_15"))
-        self.label_74 = QtGui.QLabel(self.NumberComponents)
-        self.label_74.setObjectName(_fromUtf8("label_74"))
-        self.horizontalLayout_15.addWidget(self.label_74)
-        self.horizontalLayout_16 = QtGui.QHBoxLayout()
-        self.horizontalLayout_16.setMargin(11)
-        self.horizontalLayout_16.setSpacing(6)
-        self.horizontalLayout_16.setObjectName(_fromUtf8("horizontalLayout_16"))
-        self.label_75 = QtGui.QLabel(self.NumberComponents)
-        self.label_75.setObjectName(_fromUtf8("label_75"))
-        self.horizontalLayout_16.addWidget(self.label_75)
-        self.label_76 = QtGui.QLabel(self.NumberComponents)
-        self.label_76.setObjectName(_fromUtf8("label_76"))
-        self.horizontalLayout_16.addWidget(self.label_76)
-        self.label_77 = QtGui.QLabel(self.NumberComponents)
-        self.label_77.setObjectName(_fromUtf8("label_77"))
-        self.horizontalLayout_16.addWidget(self.label_77)
-        self.label_78 = QtGui.QLabel(self.NumberComponents)
-        self.label_78.setObjectName(_fromUtf8("label_78"))
-        self.horizontalLayout_16.addWidget(self.label_78)
-        self.label_79 = QtGui.QLabel(self.NumberComponents)
-        self.label_79.setObjectName(_fromUtf8("label_79"))
-        self.horizontalLayout_16.addWidget(self.label_79)
-        self.label_83 = QtGui.QLabel(self.NumberComponents)
-        self.label_83.setObjectName(_fromUtf8("label_83"))
-        self.horizontalLayout_16.addWidget(self.label_83)
-        self.label_84 = QtGui.QLabel(self.NumberComponents)
-        self.label_84.setObjectName(_fromUtf8("label_84"))
-        self.horizontalLayout_16.addWidget(self.label_84)
-        self.label_88 = QtGui.QLabel(self.NumberComponents)
-        self.label_88.setObjectName(_fromUtf8("label_88"))
-        self.horizontalLayout_16.addWidget(self.label_88)
-        self.label_89 = QtGui.QLabel(self.NumberComponents)
-        self.label_89.setObjectName(_fromUtf8("label_89"))
-        self.horizontalLayout_16.addWidget(self.label_89)
-        self.label_90 = QtGui.QLabel(self.NumberComponents)
-        self.label_90.setObjectName(_fromUtf8("label_90"))
-        self.horizontalLayout_16.addWidget(self.label_90)
-        self.label_91 = QtGui.QLabel(self.NumberComponents)
-        self.label_91.setObjectName(_fromUtf8("label_91"))
-        self.horizontalLayout_16.addWidget(self.label_91)
-        self.horizontalLayout_15.addLayout(self.horizontalLayout_16)
-        self.verticalLayout_17.addLayout(self.horizontalLayout_15)
-        self.verticalLayout_8.addWidget(self.NumberComponents)
-        self.NumberComponents.raise_()
-        QtCore.QObject.connect(self.create_model_spin_2, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")),
-                               self.label_79.setNum)
-        QtCore.QObject.connect(self.create_model_spin, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")),
-                               self.label_77.setNum)
-        QtCore.QObject.connect(self.create_model_spin_3, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")),
-                               self.label_84.setNum)
-        QtCore.QObject.connect(self.create_model_spin_4, QtCore.SIGNAL(_fromUtf8("valueChanged(int)")),
-                               self.label_89.setNum)
-        self.NumberComponents.setTitle(_translate("MainWindow", "Number of Components", None))
-        self.label_70.setText(_translate("MainWindow", "Value 1", None))
-        self.label_71.setText(_translate("MainWindow", "Value 2", None))
-        self.label_72.setText(_translate("MainWindow", "Value 3", None))
-        self.label_73.setText(_translate("MainWindow", "Value 4", None))
-        self.label_74.setText(_translate("MainWindow", "ncs", None))
-        self.label_75.setText(_translate("MainWindow", "[", None))
-        self.label_76.setText(_translate("MainWindow", "(", None))
-        self.label_77.setText(_translate("MainWindow", "0", None))
-        self.label_78.setText(_translate("MainWindow", ",", None))
-        self.label_79.setText(_translate("MainWindow", "0", None))
-        self.label_83.setText(_translate("MainWindow", ",", None))
-        self.label_84.setText(_translate("MainWindow", "0", None))
-        self.label_88.setText(_translate("MainWindow", ",", None))
-        self.label_89.setText(_translate("MainWindow", "0", None))
-        self.label_90.setText(_translate("MainWindow", ")", None))
-        self.label_91.setText(_translate("MainWindow", "]", None))
-        self.num_of_components_button = QtGui.QPushButton(self.NumberComponents)
-        self.set_module_button(self.verticalLayout_17, self.num_of_components_button)
-        try:
-            self.num_of_components_button.clicked.connect(
-                lambda: self.pysat_fun.get_number_components([int(self.create_model_spin.text()),
-                                                              int(self.create_model_spin_2.text()),
-                                                              int(self.create_model_spin_3.text()),
-                                                              int(self.create_model_spin_4.text())]))
-        except:
-            pass
-
-    def y_variable(self, MainWindow):
-        self.y_variable = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.y_variable.setFont(font)
-        self.y_variable.setObjectName(_fromUtf8("y_variable"))
-        self.verticalLayout_25 = QtGui.QVBoxLayout(self.y_variable)
-        self.verticalLayout_25.setMargin(11)
-        self.verticalLayout_25.setSpacing(6)
-        self.verticalLayout_25.setObjectName(_fromUtf8("CrossValid"))
-        self.horizontalLayout_4 = QtGui.QHBoxLayout()
-        self.horizontalLayout_4.setMargin(11)
-        self.horizontalLayout_4.setSpacing(6)
-        self.horizontalLayout_4.setObjectName(_fromUtf8("horizontalLayout_4"))
-        self.verticalLayout_11 = QtGui.QVBoxLayout()
-        self.verticalLayout_11.setMargin(11)
-        self.verticalLayout_11.setSpacing(6)
-        self.verticalLayout_11.setObjectName(_fromUtf8("verticalLayout_11"))
-        self.label_14 = QtGui.QLabel(self.y_variable)
-        self.label_14.setObjectName(_fromUtf8("label_14"))
-        self.verticalLayout_11.addWidget(self.label_14)
-        self.horizontalLayout_4.addLayout(self.verticalLayout_11)
-        self.verticalLayout_12 = QtGui.QVBoxLayout()
-        self.verticalLayout_12.setMargin(11)
-        self.verticalLayout_12.setSpacing(6)
-        self.verticalLayout_12.setObjectName(_fromUtf8("verticalLayout_12"))
-        self.elementNameLine = QtGui.QLineEdit(self.y_variable)
-        self.elementNameLine.setObjectName(_fromUtf8("elementNameLine"))
-        self.verticalLayout_12.addWidget(self.elementNameLine)
-        self.horizontalLayout_4.addLayout(self.verticalLayout_12)
-        self.verticalLayout_25.addLayout(self.horizontalLayout_4)
-        self.verticalLayout_8.addWidget(self.y_variable)
-        self.y_variable.raise_()
-        self.y_variable.setTitle(_translate("MainWindow", "Y Variable", None))
-        self.label_14.setText(_translate("MainWindow", "Element Name", None))
-        self.y_variable_button = QtGui.QPushButton(self.y_variable)
-        self.set_module_button(self.verticalLayout_25, self.y_variable_button)
-        try:
-            self.y_variable_button.clicked.connect(lambda: self.pysat_fun.set_element_name(self.elementNameLine.text()))
-        except:
-            pass
-
-    def interpolated(self, MainWindow):
-        self.Interpolated = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.Interpolated.setFont(font)
-        self.Interpolated.setObjectName(_fromUtf8("Interpolated"))
-        self.verticalLayout_29 = QtGui.QVBoxLayout(self.Interpolated)
-        self.verticalLayout_29.setMargin(11)
-        self.verticalLayout_29.setSpacing(6)
-        self.verticalLayout_29.setObjectName(_fromUtf8("verticalLayout_29"))
-        self.horizontalLayout_12 = QtGui.QHBoxLayout()
-        self.horizontalLayout_12.setMargin(11)
-        self.horizontalLayout_12.setSpacing(6)
-        self.horizontalLayout_12.setObjectName(_fromUtf8("horizontalLayout_12"))
-        self.verticalLayout_23 = QtGui.QVBoxLayout()
-        self.verticalLayout_23.setMargin(11)
-        self.verticalLayout_23.setSpacing(6)
-        self.verticalLayout_23.setObjectName(_fromUtf8("verticalLayout_23"))
-        self.label_16 = QtGui.QLabel(self.Interpolated)
-        self.label_16.setObjectName(_fromUtf8("label_16"))
-        self.verticalLayout_23.addWidget(self.label_16)
-        self.horizontalLayout_12.addLayout(self.verticalLayout_23)
-        self.verticalLayout_29.addLayout(self.horizontalLayout_12)
-        self.verticalLayout_8.addWidget(self.Interpolated)
-        self.Interpolated.setTitle(_translate("MainWindow", "Interpolated", None))
-        self.label_16.setText(_translate("MainWindow", "Interpolated unknown data to known data", None))
-        self.pysat_fun.set_interp()
-
-    def masked(self, MainWindow):
-        self.masked = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.masked.setFont(font)
-        self.masked.setObjectName(_fromUtf8("masked"))
-        self.verticalLayout_2 = QtGui.QVBoxLayout(self.masked)
-        self.verticalLayout_2.setMargin(11)
-        self.verticalLayout_2.setSpacing(6)
-        self.verticalLayout_2.setObjectName(_fromUtf8("verticalLayout_2"))
-        self.verticalLayout = QtGui.QVBoxLayout()
-        self.verticalLayout.setMargin(11)
-        self.verticalLayout.setSpacing(6)
-        self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
-        self.horizontalLayout_13 = QtGui.QHBoxLayout()
-        self.horizontalLayout_13.setMargin(11)
-        self.horizontalLayout_13.setSpacing(6)
-        self.horizontalLayout_13.setObjectName(_fromUtf8("horizontalLayout_13"))
-        self.verticalLayout_17 = QtGui.QVBoxLayout()
-        self.verticalLayout_17.setMargin(11)
-        self.verticalLayout_17.setSpacing(6)
-        self.verticalLayout_17.setObjectName(_fromUtf8("verticalLayout_17"))
-        self.label_9 = QtGui.QLabel(self.masked)
-        self.label_9.setObjectName(_fromUtf8("label_9"))
-        self.verticalLayout_17.addWidget(self.label_9)
-        self.horizontalLayout_13.addLayout(self.verticalLayout_17)
-        self.verticalLayout_22 = QtGui.QVBoxLayout()
-        self.verticalLayout_22.setMargin(11)
-        self.verticalLayout_22.setSpacing(6)
-        self.verticalLayout_22.setObjectName(_fromUtf8("verticalLayout_22"))
-        self.lineEdit_10 = QtGui.QLineEdit(self.masked)
-        self.lineEdit_10.setReadOnly(True)
-        self.lineEdit_10.setObjectName(_fromUtf8("lineEdit_10"))
-        self.verticalLayout_22.addWidget(self.lineEdit_10)
-        self.horizontalLayout_13.addLayout(self.verticalLayout_22)
-        self.verticalLayout_24 = QtGui.QVBoxLayout()
-        self.verticalLayout_24.setMargin(11)
-        self.verticalLayout_24.setSpacing(6)
-        self.verticalLayout_24.setObjectName(_fromUtf8("verticalLayout_24"))
-        self.fullDataBaseButton_4 = QtGui.QToolButton(self.masked)
-        self.fullDataBaseButton_4.setObjectName(_fromUtf8("fullDataBaseButton_4"))
-        self.verticalLayout_24.addWidget(self.fullDataBaseButton_4)
-        self.horizontalLayout_13.addLayout(self.verticalLayout_24)
-        self.verticalLayout.addLayout(self.horizontalLayout_13)
-        self.label_16 = QtGui.QLabel(self.masked)
-        self.label_16.setObjectName(_fromUtf8("label_16"))
-        self.verticalLayout.addWidget(self.label_16)
-        self.verticalLayout_2.addLayout(self.verticalLayout)
-        self.verticalLayout_8.addWidget(self.masked)
-        self.masked.setTitle(_translate("MainWindow", "Mask", None))
-        self.label_9.setText(_translate("MainWindow", "Maskfile", None))
-        self.lineEdit_10.setText(_translate("MainWindow", "*.csv", None))
-        self.fullDataBaseButton_4.setText(_translate("MainWindow", "...", None))
-        self.label_16.setText(_translate("MainWindow", "Mask was applied to unknown and known data", None))
-        try:
-            self.fullDataBaseButton_4.clicked.connect(lambda: pysat_ui.on_maskFile_clicked(self, self.lineEdit_10))
-        except:
-            pass
-
-
 
 
 
@@ -611,7 +22,7 @@
     def on_outPutLocationButton_clicked(self, lineEdit):
         filename = QFileDialog.getExistingDirectory(None, "Select Output Directory", '.')
         lineEdit.setText(filename)
-        self.pysat_fun.set_file_outpath(filename)
+        self.pysat_fun.file_outpath(filename)
         if lineEdit.text() == "":
             lineEdit.setText("*/*")
 
@@ -741,3 +152,196 @@ return listwidget
 return listwidget
 >> >> >> > 18
 fa50f900de0cff5a4d52115c9310b4d63de617
+
+
+
+    def stratified_folds(self, MainWindow):
+        self.strat_folds = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.strat_folds.setFont(font)
+        self.strat_folds.setObjectName(_fromUtf8("Stratified Folds"))
+        self.strat_folds_vlayout = QtGui.QVBoxLayout(self.strat_folds)
+        self.strat_folds_vlayout.setObjectName(_fromUtf8("strat_folds_vlayout"))
+        self.strat_folds_choose_data_label = QtGui.QLabel(self.strat_folds)
+        self.strat_folds_choose_data_label.setObjectName(_fromUtf8("strat_folds_choose_data_label"))
+        self.strat_folds_vlayout.addWidget(self.strat_folds_choose_data_label)
+        datachoices = self.pysat_fun.datakeys
+        if datachoices == []:
+            datachoices = ['No data has been loaded!']
+        self.strat_folds_choose_data = make_combobox(datachoices)
+        self.strat_folds_vlayout.addWidget(self.strat_folds_choose_data)
+        self.strat_folds_choose_var_label = QtGui.QLabel(self.strat_folds)
+        self.strat_folds_choose_var_label.setObjectName(_fromUtf8("strat_folds_choose_var_label"))
+        self.strat_folds_vlayout.addWidget(self.strat_folds_choose_var_label)
+
+        varchoices = self.pysat_fun.data[self.strat_folds_choose_data.currentText()].df['meta'].columns.values
+        self.strat_folds_choose_var = make_combobox(varchoices)
+        self.strat_folds_vlayout.addWidget(self.strat_folds_choose_var)
+        self.strat_folds_choose_data.activated[int].connect(self.strat_fold_change_vars)
+        self.strat_folds_hlayout = QtGui.QHBoxLayout()
+        self.strat_folds_hlayout.setObjectName(_fromUtf8("strat_folds_hlayout"))
+        self.nfolds_label = QtGui.QLabel(self.strat_folds)
+        self.nfolds_label.setObjectName(_fromUtf8("nfolds_label"))
+        self.strat_folds_hlayout.addWidget(self.nfolds_label)
+        self.nfolds_spin = QtGui.QSpinBox(self.strat_folds)
+        self.nfolds_spin.setObjectName(_fromUtf8("nfolds_spin"))
+        self.nfolds_spin.setMinimum(1)
+        self.strat_folds_hlayout.addWidget(self.nfolds_spin)
+        self.test_fold_label = QtGui.QLabel(self.strat_folds)
+        self.test_fold_label.setObjectName(_fromUtf8("test_fold_label"))
+        self.strat_folds_hlayout.addWidget(self.test_fold_label)
+        foldchoices = ['1']
+        self.choose_test_fold = make_combobox(foldchoices)
+        self.choose_test_fold.setObjectName(_fromUtf8("choose_test_fold"))
+        self.nfolds_spin.valueChanged.connect(self.strat_fold_change_testfolds)
+        self.strat_folds_hlayout.addWidget(self.choose_test_fold)
+        self.spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
+        self.strat_folds_hlayout.addItem(self.spacerItem)
+        self.create_folds = QtGui.QPushButton(self.strat_folds)
+        self.create_folds.setObjectName(_fromUtf8("create_folds"))
+        self.create_folds.setText(_translate("strat_folds", "Create Folds", None))
+        self.strat_folds_hlayout.addWidget(self.create_folds)
+        self.strat_folds_vlayout.addLayout(self.strat_folds_hlayout)
+        self.verticalLayout_8.addWidget(self.strat_folds)
+        self.strat_folds.raise_()
+        self.strat_folds.setTitle(_translate("MainWindow", "Stratified Folds", None))
+        self.nfolds_label.setText(_translate("strat_folds", "N folds", None))
+        self.test_fold_label.setText(_translate("strat_folds", "Test Fold", None))
+        self.create_folds.setText(_translate("strat_folds", "Create Folds", None))
+        self.strat_folds_choose_data_label.setText(_translate("strat_folds", "Choose data to stratify:", None))
+        self.strat_folds_choose_var_label.setText(
+            _translate("strat_folds", "Choose variable on which to sort:", None))
+        try:
+            self.create_folds.clicked.connect(
+                lambda: self.pysat_fun.do_strat_folds(datakey=str(self.strat_folds_choose_data.currentText()),
+                                                      nfolds=int(self.nfolds_spin.text()),
+                                                      testfold=int(self.choose_test_fold.currentText()),
+                                                      colname=('comp', self.strat_folds_choose_var.currentText())))
+        except:
+            print('There was a problem with creating stratified folds...')
+
+
+            ####These functions below are private and add functionality to the UI
+
+    def submodel(self, MainWindow):
+        pass
+
+    def masked(self, MainWindow):
+        self.masked = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
+        font = QtGui.QFont()
+        font.setPointSize(10)
+        self.masked.setFont(font)
+        self.masked.setObjectName(_fromUtf8("masked"))
+        self.verticalLayout_2 = QtGui.QVBoxLayout(self.masked)
+        self.verticalLayout_2.setMargin(11)
+        self.verticalLayout_2.setSpacing(6)
+        self.verticalLayout_2.setObjectName(_fromUtf8("verticalLayout_2"))
+        self.verticalLayout = QtGui.QVBoxLayout()
+        self.verticalLayout.setMargin(11)
+        self.verticalLayout.setSpacing(6)
+        self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
+        self.horizontalLayout_13 = QtGui.QHBoxLayout()
+        self.horizontalLayout_13.setMargin(11)
+        self.horizontalLayout_13.setSpacing(6)
+        self.horizontalLayout_13.setObjectName(_fromUtf8("horizontalLayout_13"))
+        self.verticalLayout_17 = QtGui.QVBoxLayout()
+        self.verticalLayout_17.setMargin(11)
+        self.verticalLayout_17.setSpacing(6)
+        self.verticalLayout_17.setObjectName(_fromUtf8("verticalLayout_17"))
+        self.label_9 = QtGui.QLabel(self.masked)
+        self.label_9.setObjectName(_fromUtf8("label_9"))
+        self.verticalLayout_17.addWidget(self.label_9)
+        self.horizontalLayout_13.addLayout(self.verticalLayout_17)
+        self.verticalLayout_22 = QtGui.QVBoxLayout()
+        self.verticalLayout_22.setMargin(11)
+        self.verticalLayout_22.setSpacing(6)
+        self.verticalLayout_22.setObjectName(_fromUtf8("verticalLayout_22"))
+        self.lineEdit_10 = QtGui.QLineEdit(self.masked)
+        self.lineEdit_10.setReadOnly(True)
+        self.lineEdit_10.setObjectName(_fromUtf8("lineEdit_10"))
+        self.verticalLayout_22.addWidget(self.lineEdit_10)
+        self.horizontalLayout_13.addLayout(self.verticalLayout_22)
+        self.verticalLayout_24 = QtGui.QVBoxLayout()
+        self.verticalLayout_24.setMargin(11)
+        self.verticalLayout_24.setSpacing(6)
+        self.verticalLayout_24.setObjectName(_fromUtf8("verticalLayout_24"))
+        self.fullDataBaseButton_4 = QtGui.QToolButton(self.masked)
+        self.fullDataBaseButton_4.setObjectName(_fromUtf8("fullDataBaseButton_4"))
+        self.verticalLayout_24.addWidget(self.fullDataBaseButton_4)
+        self.horizontalLayout_13.addLayout(self.verticalLayout_24)
+        self.verticalLayout.addLayout(self.horizontalLayout_13)
+        self.label_16 = QtGui.QLabel(self.masked)
+        self.label_16.setObjectName(_fromUtf8("label_16"))
+        self.verticalLayout.addWidget(self.label_16)
+        self.verticalLayout_2.addLayout(self.verticalLayout)
+        self.verticalLayout_8.addWidget(self.masked)
+        self.masked.setTitle(_translate("MainWindow", "Mask", None))
+        self.label_9.setText(_translate("MainWindow", "Maskfile", None))
+        self.lineEdit_10.setText(_translate("MainWindow", "*.csv", None))
+        self.fullDataBaseButton_4.setText(_translate("MainWindow", "...", None))
+        self.label_16.setText(_translate("MainWindow", "Mask was applied to unknown and known data", None))
+        try:
+            self.fullDataBaseButton_4.clicked.connect(lambda: pysat_ui.on_maskFile_clicked(self, self.lineEdit_10))
+        except:
+            pass
+
+    """ =============================================
+    Please do not delete the files below this line!
+    These files are the working files that allow the UI
+    to operate and do things!
+    ============================================== """
+
+    def set_module_button(self, QVBoxLayout, PushButtonName):
+        self.horizontalLayout = QtGui.QHBoxLayout()
+        self.horizontalLayout.setMargin(11)
+        self.horizontalLayout.setSpacing(6)
+        spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
+        self.horizontalLayout.addItem(spacerItem)
+        PushButtonName.setObjectName(_fromUtf8("run_module_files"))
+        self.horizontalLayout.addWidget(PushButtonName)
+        QVBoxLayout.addLayout(self.horizontalLayout)
+        PushButtonName.setText(_translate("MainWindow", "Run Module", None))
+
+    def menu_item_shortcuts(self):
+        self.actionExit.setShortcut("ctrl+Q")
+
+    def menu_item_functions(self, MainWindow):
+        self.actionSet_output_location.triggered.connect(lambda: pysat_ui.file_outpath(self, MainWindow))
+        self.actionLoad_Unknown_Data.triggered.connect(lambda: pysat_ui.unknown_data(self, MainWindow))
+        self.actionLoad_reference_Data.triggered.connect(lambda: pysat_ui.reference_data(self, MainWindow))
+        self.actionNormalization.triggered.connect(lambda: pysat_ui.normalization(self, MainWindow))
+        self.actionInterpolate.triggered.connect(lambda: pysat_ui.interpolated(self, MainWindow))
+        self.actionApply_Mask.triggered.connect(lambda: pysat_ui.masked(self, MainWindow))
+        self.actionStratified_Folds.triggered.connect(lambda: pysat_ui.stratified_folds(self, MainWindow))
+        self.actionTrain.triggered.connect(lambda: pysat_ui.regression_train(self, MainWindow))
+
+    def saveworkflow(self):
+        # TODO save the current window's data into a save file
+        pass
+
+    def openworkflow(self):
+        # TODO open file dialog
+        self.filename = QFileDialog.getOpenFileName(self, "Open a Workflow File", '.', "(*.wrf)")
+
+    def on_maskFile_clicked(self, lineEdit):
+        filename = QFileDialog.getOpenFileName(None, "Open Mask File", '.', "(*.csv)")
+        lineEdit.setText(filename)
+        self.pysat_fun.set_file_maskfile(filename)
+        if lineEdit.text() == "":
+            lineEdit.setText("*.csv")
+        self.pysat_fun.set_mask()
+
+    def on_getDataButton_clicked(self, lineEdit, key):
+        filename = QFileDialog.getOpenFileName(None, "Open Uknown Data File", '.', "(*.csv)")
+        lineEdit.setText(filename)
+        if lineEdit.text() == "":
+            lineEdit.setText("*.csv")
+        self.pysat_fun.get_data(filename, key)
+
+    def on_okButton_clicked(self):
+        self.pysat_fun.set_sm()
+        self.pysat_fun.get_sm_fit()
+        self.pysat_fun.get_plots()
+
+
