--- src/Python PYSAT_GUI/PythonUI/mainwindow_regression.py
@@ -1,232 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file 'C:\Users\nfinch\Desktop\GitHub\PySAT\src\PYSAT_Gui_UI_Forms\01_mainwindow_regression.ui'
-#
-# Created by: PyQt4 UI code generator 4.11.4
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt4 import QtCore, QtGui
-
-try:
-    _fromUtf8 = QtCore.QString.fromUtf8
-except AttributeError:
-    def _fromUtf8(s):
-        return s
-
-try:
-    _encoding = QtGui.QApplication.UnicodeUTF8
-    def _translate(context, text, disambig):
-        return QtGui.QApplication.translate(context, text, disambig, _encoding)
-except AttributeError:
-    def _translate(context, text, disambig):
-        return QtGui.QApplication.translate(context, text, disambig)
-
-class Ui_regression(object):
-    def setupUi(self, regression):
-        regression.setObjectName(_fromUtf8("regression"))
-        regression.resize(644, 525)
-        regression.setMinimumSize(QtCore.QSize(600, 150))
-        regression.setMaximumSize(QtCore.QSize(16777215, 999999))
-        self.verticalLayout_2 = QtGui.QVBoxLayout(regression)
-        self.verticalLayout_2.setObjectName(_fromUtf8("verticalLayout_2"))
-        self.regression_vlayout = QtGui.QVBoxLayout()
-        self.regression_vlayout.setObjectName(_fromUtf8("regression_vlayout"))
-        self.regression_choosedata_hlayout = QtGui.QHBoxLayout()
-        self.regression_choosedata_hlayout.setObjectName(_fromUtf8("regression_choosedata_hlayout"))
-        self.label = QtGui.QLabel(regression)
-        self.label.setObjectName(_fromUtf8("label"))
-        self.regression_choosedata_hlayout.addWidget(self.label)
-        self.regression_choosedata = QtGui.QComboBox(regression)
-        self.regression_choosedata.setIconSize(QtCore.QSize(50, 20))
-        self.regression_choosedata.setObjectName(_fromUtf8("regression_choosedata"))
-        self.regression_choosedata.addItem(_fromUtf8(""))
-        self.regression_choosedata.addItem(_fromUtf8(""))
-        self.regression_choosedata_hlayout.addWidget(self.regression_choosedata)
-        spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.regression_choosedata_hlayout.addItem(spacerItem)
-        self.regression_vlayout.addLayout(self.regression_choosedata_hlayout)
-        self.regression_choosevars_hlayout = QtGui.QHBoxLayout()
-        self.regression_choosevars_hlayout.setObjectName(_fromUtf8("regression_choosevars_hlayout"))
-        self.regression_train_choosex_label = QtGui.QLabel(regression)
-        self.regression_train_choosex_label.setObjectName(_fromUtf8("regression_train_choosex_label"))
-        self.regression_choosevars_hlayout.addWidget(self.regression_train_choosex_label)
-        self.regression_train_choosex = QtGui.QListWidget(regression)
-        self.regression_train_choosex.setObjectName(_fromUtf8("regression_train_choosex"))
-        item = QtGui.QListWidgetItem()
-        self.regression_train_choosex.addItem(item)
-        self.regression_choosevars_hlayout.addWidget(self.regression_train_choosex)
-        self.regression_train_choosey_label = QtGui.QLabel(regression)
-        self.regression_train_choosey_label.setObjectName(_fromUtf8("regression_train_choosey_label"))
-        self.regression_choosevars_hlayout.addWidget(self.regression_train_choosey_label)
-        self.regression_train_choosey = QtGui.QListWidget(regression)
-        self.regression_train_choosey.setObjectName(_fromUtf8("regression_train_choosey"))
-        item = QtGui.QListWidgetItem()
-        self.regression_train_choosey.addItem(item)
-        self.regression_choosevars_hlayout.addWidget(self.regression_train_choosey)
-        spacerItem1 = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.regression_choosevars_hlayout.addItem(spacerItem1)
-        self.regression_vlayout.addLayout(self.regression_choosevars_hlayout)
-        self.regression_choosealg_hlayout = QtGui.QHBoxLayout()
-        self.regression_choosealg_hlayout.setObjectName(_fromUtf8("regression_choosealg_hlayout"))
-        self.regression_choosealg_label = QtGui.QLabel(regression)
-        self.regression_choosealg_label.setObjectName(_fromUtf8("regression_choosealg_label"))
-        self.regression_choosealg_hlayout.addWidget(self.regression_choosealg_label)
-        self.regression_choosealg = QtGui.QComboBox(regression)
-        self.regression_choosealg.setIconSize(QtCore.QSize(50, 20))
-        self.regression_choosealg.setObjectName(_fromUtf8("regression_choosealg"))
-        self.regression_choosealg.addItem(_fromUtf8(""))
-        self.regression_choosealg.addItem(_fromUtf8(""))
-        self.regression_choosealg.addItem(_fromUtf8(""))
-        self.regression_choosealg_hlayout.addWidget(self.regression_choosealg)
-        spacerItem2 = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.regression_choosealg_hlayout.addItem(spacerItem2)
-        self.regression_vlayout.addLayout(self.regression_choosealg_hlayout)
-        self.pls_widget = QtGui.QWidget(regression)
-        self.pls_widget.setMinimumSize(QtCore.QSize(0, 0))
-        self.pls_widget.setObjectName(_fromUtf8("pls_widget"))
-        self.horizontalLayout_2 = QtGui.QHBoxLayout(self.pls_widget)
-        self.horizontalLayout_2.setObjectName(_fromUtf8("horizontalLayout_2"))
-        self.pls_nc_label = QtGui.QLabel(self.pls_widget)
-        self.pls_nc_label.setObjectName(_fromUtf8("pls_nc_label"))
-        self.horizontalLayout_2.addWidget(self.pls_nc_label)
-        self.pls_nc_spinbox = QtGui.QSpinBox(self.pls_widget)
-        self.pls_nc_spinbox.setObjectName(_fromUtf8("pls_nc_spinbox"))
-        self.horizontalLayout_2.addWidget(self.pls_nc_spinbox)
-        spacerItem3 = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.horizontalLayout_2.addItem(spacerItem3)
-        self.regression_vlayout.addWidget(self.pls_widget)
-        self.gp_widget = QtGui.QWidget(regression)
-        self.gp_widget.setObjectName(_fromUtf8("gp_widget"))
-        self.gp_vlayout = QtGui.QVBoxLayout(self.gp_widget)
-        self.gp_vlayout.setObjectName(_fromUtf8("gp_vlayout"))
-        self.gp_dim_red_hlayout = QtGui.QHBoxLayout()
-        self.gp_dim_red_hlayout.setObjectName(_fromUtf8("gp_dim_red_hlayout"))
-        self.gp_dim_red_label = QtGui.QLabel(self.gp_widget)
-        self.gp_dim_red_label.setObjectName(_fromUtf8("gp_dim_red_label"))
-        self.gp_dim_red_hlayout.addWidget(self.gp_dim_red_label)
-        self.gp_dim_red_combobox = QtGui.QComboBox(self.gp_widget)
-        self.gp_dim_red_combobox.setObjectName(_fromUtf8("gp_dim_red_combobox"))
-        self.gp_dim_red_combobox.addItem(_fromUtf8(""))
-        self.gp_dim_red_combobox.addItem(_fromUtf8(""))
-        self.gp_dim_red_combobox.addItem(_fromUtf8(""))
-        self.gp_dim_red_combobox.addItem(_fromUtf8(""))
-        self.gp_dim_red_hlayout.addWidget(self.gp_dim_red_combobox)
-        self.gp_vlayout.addLayout(self.gp_dim_red_hlayout)
-        self.gp_rand_starts_hlayout = QtGui.QHBoxLayout()
-        self.gp_rand_starts_hlayout.setObjectName(_fromUtf8("gp_rand_starts_hlayout"))
-        self.gp_rand_starts_label = QtGui.QLabel(self.gp_widget)
-        self.gp_rand_starts_label.setObjectName(_fromUtf8("gp_rand_starts_label"))
-        self.gp_rand_starts_hlayout.addWidget(self.gp_rand_starts_label)
-        self.gp_rand_starts_spin = QtGui.QSpinBox(self.gp_widget)
-        self.gp_rand_starts_spin.setObjectName(_fromUtf8("gp_rand_starts_spin"))
-        self.gp_rand_starts_hlayout.addWidget(self.gp_rand_starts_spin)
-        spacerItem4 = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.gp_rand_starts_hlayout.addItem(spacerItem4)
-        self.gp_vlayout.addLayout(self.gp_rand_starts_hlayout)
-        self.gp_theta_hlayout = QtGui.QHBoxLayout()
-        self.gp_theta_hlayout.setObjectName(_fromUtf8("gp_theta_hlayout"))
-        self.gp_theta0_label = QtGui.QLabel(self.gp_widget)
-        self.gp_theta0_label.setObjectName(_fromUtf8("gp_theta0_label"))
-        self.gp_theta_hlayout.addWidget(self.gp_theta0_label)
-        self.gp_theta0_spin = QtGui.QDoubleSpinBox(self.gp_widget)
-        self.gp_theta0_spin.setObjectName(_fromUtf8("gp_theta0_spin"))
-        self.gp_theta_hlayout.addWidget(self.gp_theta0_spin)
-        self.gp_thetaL_label = QtGui.QLabel(self.gp_widget)
-        self.gp_thetaL_label.setObjectName(_fromUtf8("gp_thetaL_label"))
-        self.gp_theta_hlayout.addWidget(self.gp_thetaL_label)
-        self.gp_thetaL_spin = QtGui.QDoubleSpinBox(self.gp_widget)
-        self.gp_thetaL_spin.setObjectName(_fromUtf8("gp_thetaL_spin"))
-        self.gp_theta_hlayout.addWidget(self.gp_thetaL_spin)
-        self.gp_thetaU_label = QtGui.QLabel(self.gp_widget)
-        self.gp_thetaU_label.setObjectName(_fromUtf8("gp_thetaU_label"))
-        self.gp_theta_hlayout.addWidget(self.gp_thetaU_label)
-        self.gp_thetaU_spin = QtGui.QDoubleSpinBox(self.gp_widget)
-        self.gp_thetaU_spin.setObjectName(_fromUtf8("gp_thetaU_spin"))
-        self.gp_theta_hlayout.addWidget(self.gp_thetaU_spin)
-        spacerItem5 = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.gp_theta_hlayout.addItem(spacerItem5)
-        self.gp_vlayout.addLayout(self.gp_theta_hlayout)
-        self.regression_vlayout.addWidget(self.gp_widget)
-        self.regression_ransac_checkbox = QtGui.QCheckBox(regression)
-        self.regression_ransac_checkbox.setObjectName(_fromUtf8("regression_ransac_checkbox"))
-        self.regression_vlayout.addWidget(self.regression_ransac_checkbox)
-        self.ransac_vlayout = QtGui.QVBoxLayout()
-        self.ransac_vlayout.setObjectName(_fromUtf8("ransac_vlayout"))
-        self.ransac_lossfunc_hlayout = QtGui.QHBoxLayout()
-        self.ransac_lossfunc_hlayout.setObjectName(_fromUtf8("ransac_lossfunc_hlayout"))
-        self.ransac_lossfunc_label = QtGui.QLabel(regression)
-        self.ransac_lossfunc_label.setObjectName(_fromUtf8("ransac_lossfunc_label"))
-        self.ransac_lossfunc_hlayout.addWidget(self.ransac_lossfunc_label)
-        self.ransac_lossfunc_combobox = QtGui.QComboBox(regression)
-        self.ransac_lossfunc_combobox.setObjectName(_fromUtf8("ransac_lossfunc_combobox"))
-        self.ransac_lossfunc_combobox.addItem(_fromUtf8(""))
-        self.ransac_lossfunc_combobox.addItem(_fromUtf8(""))
-        self.ransac_lossfunc_hlayout.addWidget(self.ransac_lossfunc_combobox)
-        spacerItem6 = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.ransac_lossfunc_hlayout.addItem(spacerItem6)
-        self.ransac_vlayout.addLayout(self.ransac_lossfunc_hlayout)
-        self.ransac_thresh_hlayout = QtGui.QHBoxLayout()
-        self.ransac_thresh_hlayout.setObjectName(_fromUtf8("ransac_thresh_hlayout"))
-        self.ransac_thresh_label = QtGui.QLabel(regression)
-        self.ransac_thresh_label.setObjectName(_fromUtf8("ransac_thresh_label"))
-        self.ransac_thresh_hlayout.addWidget(self.ransac_thresh_label)
-        self.ransac_thresh_spin = QtGui.QDoubleSpinBox(regression)
-        self.ransac_thresh_spin.setObjectName(_fromUtf8("ransac_thresh_spin"))
-        self.ransac_thresh_hlayout.addWidget(self.ransac_thresh_spin)
-        spacerItem7 = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.ransac_thresh_hlayout.addItem(spacerItem7)
-        self.ransac_vlayout.addLayout(self.ransac_thresh_hlayout)
-        self.regression_vlayout.addLayout(self.ransac_vlayout)
-        self.regression_trainbutton_hlayout = QtGui.QHBoxLayout()
-        self.regression_trainbutton_hlayout.setObjectName(_fromUtf8("regression_trainbutton_hlayout"))
-        spacerItem8 = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.regression_trainbutton_hlayout.addItem(spacerItem8)
-        self.regression_trainbutton = QtGui.QPushButton(regression)
-        self.regression_trainbutton.setObjectName(_fromUtf8("regression_trainbutton"))
-        self.regression_trainbutton_hlayout.addWidget(self.regression_trainbutton)
-        self.regression_vlayout.addLayout(self.regression_trainbutton_hlayout)
-        self.verticalLayout_2.addLayout(self.regression_vlayout)
-
-        self.retranslateUi(regression)
-        QtCore.QMetaObject.connectSlotsByName(regression)
-
-    def retranslateUi(self, regression):
-        regression.setWindowTitle(_translate("regression", "Regression", None))
-        regression.setTitle(_translate("regression", "Regression", None))
-        self.label.setText(_translate("regression", "Choose data: ", None))
-        self.regression_choosedata.setItemText(0, _translate("regression", "Choose Data", None))
-        self.regression_choosedata.setItemText(1, _translate("regression", "Known Data", None))
-        self.regression_train_choosex_label.setText(_translate("regression", "Choose X variable(s): ", None))
-        __sortingEnabled = self.regression_train_choosex.isSortingEnabled()
-        self.regression_train_choosex.setSortingEnabled(False)
-        item = self.regression_train_choosex.item(0)
-        item.setText(_translate("regression", "Choose X", None))
-        self.regression_train_choosex.setSortingEnabled(__sortingEnabled)
-        self.regression_train_choosey_label.setText(_translate("regression", "Choose Y variable(s):", None))
-        __sortingEnabled = self.regression_train_choosey.isSortingEnabled()
-        self.regression_train_choosey.setSortingEnabled(False)
-        item = self.regression_train_choosey.item(0)
-        item.setText(_translate("regression", "Choose Y", None))
-        self.regression_train_choosey.setSortingEnabled(__sortingEnabled)
-        self.regression_choosealg_label.setText(_translate("regression", "Choose Algorithm: ", None))
-        self.regression_choosealg.setItemText(0, _translate("regression", "PLS", None))
-        self.regression_choosealg.setItemText(1, _translate("regression", "GP", None))
-        self.regression_choosealg.setItemText(2, _translate("regression", "Others coming soon...", None))
-        self.pls_nc_label.setText(_translate("regression", "# of Components: ", None))
-        self.gp_dim_red_label.setText(_translate("regression", "Dimensionality reduction method:", None))
-        self.gp_dim_red_combobox.setItemText(0, _translate("regression", "PCA", None))
-        self.gp_dim_red_combobox.setItemText(1, _translate("regression", "FastICA", None))
-        self.gp_dim_red_combobox.setItemText(2, _translate("regression", "ICA - JADE", None))
-        self.gp_dim_red_combobox.setItemText(3, _translate("regression", "None", None))
-        self.gp_rand_starts_label.setText(_translate("regression", "# of random starts:", None))
-        self.gp_theta0_label.setText(_translate("regression", "theta_0", None))
-        self.gp_thetaL_label.setText(_translate("regression", "theta_L", None))
-        self.gp_thetaU_label.setText(_translate("regression", "theta_U", None))
-        self.regression_ransac_checkbox.setText(_translate("regression", "RANSAC", None))
-        self.ransac_lossfunc_label.setText(_translate("regression", "Loss function:", None))
-        self.ransac_lossfunc_combobox.setItemText(0, _translate("regression", "Squared", None))
-        self.ransac_lossfunc_combobox.setItemText(1, _translate("regression", "Absolute", None))
-        self.ransac_thresh_label.setText(_translate("regression", "Threshold:", None))
-        self.regression_trainbutton.setText(_translate("regression", "Train", None))
-
