+++ src/Python PYSAT_GUI/pysat_main.py
@@ -12,8 +12,9 @@ class Main(QMainWindow):
     def runningFunctions(self, MainWindow):
         pysat = pysat_ui()
         pysat.mainframe(MainWindow)           # Set up the mainwindow. This is the backbone of the UI it IS REQUIRED
-        pysat.menu_item_shortcuts()           # The shortcuts for making things happen in the UI
-        pysat.menu_item_functions(MainWindow) # These are the various functions that make the UI work
+        pysat.file_outpath(MainWindow)
+        # pysat.menu_item_shortcuts()           # The shortcuts for making things happen in the UI
+        # pysat.menu_item_functions(MainWindow) # These are the various functions that make the UI work
 
 
 #### These are the triggers for exit and new
