+++ src/Python PYSAT_GUI/pysat_ui.py
@@ -4,6 +4,10 @@ import sys
 from pysat_function import pysat_func
 from functools import partial
 from PyQt4.QtGui import QMessageBox
+from PythonUI.mainframe_ import mainframe_
+from PythonUI.file_out_ import file_outpath_
+from PythonUI.get_data_ import *
+from PythonUI.regression_ import *
 
 try:
     _fromUtf8 = QtCore.QString.fromUtf8
@@ -21,7 +25,6 @@ except AttributeError:
     def _translate(context, text, disambig):
         return QtGui.QApplication.translate(context, text, disambig)
 
-
 class pysat_ui(object):
     def __init__(self):
         self.pysat_fun = pysat_func()
@@ -29,697 +32,28 @@ class pysat_ui(object):
         self.addPara = []
 
     def mainframe(self, MainWindow):
-        MainWindow.setObjectName(_fromUtf8("MainWindow"))
-        MainWindow.resize(600, 843)
-        self.centralWidget = QtGui.QWidget(MainWindow)
-        self.centralWidget.setObjectName(_fromUtf8("centralWidget"))
-        self.verticalLayout_9 = QtGui.QVBoxLayout(self.centralWidget)
-        self.verticalLayout_9.setMargin(11)
-        self.verticalLayout_9.setSpacing(6)
-        self.verticalLayout_9.setObjectName(_fromUtf8("verticalLayout_9"))
-        self.scrollArea = QtGui.QScrollArea(self.centralWidget)
-        self.scrollArea.setWidgetResizable(True)
-        self.scrollArea.setObjectName(_fromUtf8("scrollArea"))
-        self.scrollAreaWidgetContents_2 = QtGui.QWidget()
-        self.scrollAreaWidgetContents_2.setGeometry(QtCore.QRect(0, 0, 557, 695))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        self.scrollAreaWidgetContents_2.setFont(font)
-        self.scrollAreaWidgetContents_2.setStyleSheet(_fromUtf8("QGroupBox {\n"
-                                                                "  border: 2px solid gray;\n"
-                                                                "  border-radius: 6px;\n"
-                                                                "  margin-top: 0.5em;\n"
-                                                                "}\n"
-                                                                "\n"
-                                                                "QGroupBox::title {\n"
-                                                                "\n"
-                                                                "  padding-top: -14px;\n"
-                                                                "  padding-left: 8px;\n"
-                                                                "}\n"
-                                                                ""))
-        self.scrollAreaWidgetContents_2.setObjectName(_fromUtf8("scrollAreaWidgetContents_2"))
-        self.verticalLayout_8 = QtGui.QVBoxLayout(self.scrollAreaWidgetContents_2)
-        self.verticalLayout_8.setMargin(11)
-        self.verticalLayout_8.setSpacing(6)
-        self.verticalLayout_8.setObjectName(_fromUtf8("verticalLayout_8"))
-        self.scrollArea.setWidget(self.scrollAreaWidgetContents_2)
-        self.verticalLayout_9.addWidget(self.scrollArea)
-        self.OK = QtGui.QGroupBox(self.centralWidget)
-        self.OK.setObjectName(_fromUtf8("OK"))
-        self.ok = QtGui.QHBoxLayout(self.OK)
-        self.ok.setMargin(11)
-        self.ok.setSpacing(6)
-        self.ok.setObjectName(_fromUtf8("ok"))
-        spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.ok.addItem(spacerItem)
-        self.okButton = QtGui.QPushButton(self.OK)
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        self.okButton.setFont(font)
-        self.okButton.setMouseTracking(False)
-        self.okButton.setObjectName(_fromUtf8("okButton"))
-        self.ok.addWidget(self.okButton)
-        self.verticalLayout_9.addWidget(self.OK)
-        MainWindow.setCentralWidget(self.centralWidget)
-        self.mainToolBar = QtGui.QToolBar(MainWindow)
-        self.mainToolBar.setObjectName(_fromUtf8("mainToolBar"))
-        MainWindow.addToolBar(QtCore.Qt.TopToolBarArea, self.mainToolBar)
-        self.statusBar = QtGui.QStatusBar(MainWindow)
-        self.statusBar.setObjectName(_fromUtf8("statusBar"))
-        MainWindow.setStatusBar(self.statusBar)
-        self.menuBar = QtGui.QMenuBar(MainWindow)
-        self.menuBar.setGeometry(QtCore.QRect(0, 0, 581, 26))
-        self.menuBar.setObjectName(_fromUtf8("menuBar"))
-        self.menuFile = QtGui.QMenu(self.menuBar)
-        self.menuFile.setObjectName(_fromUtf8("menuFile"))
-        self.menuPreprocessing = QtGui.QMenu(self.menuBar)
-        self.menuPreprocessing.setObjectName(_fromUtf8("menuPreprocessing"))
-        self.menuBaseline_Removal = QtGui.QMenu(self.menuPreprocessing)
-        self.menuBaseline_Removal.setObjectName(_fromUtf8("menuBaseline_Removal"))
-        self.menuCalibration_Transfer = QtGui.QMenu(self.menuPreprocessing)
-        self.menuCalibration_Transfer.setObjectName(_fromUtf8("menuCalibration_Transfer"))
-        self.menuRegression = QtGui.QMenu(self.menuBar)
-        self.menuRegression.setObjectName(_fromUtf8("menuRegression"))
-        self.menuHelp = QtGui.QMenu(self.menuBar)
-        self.menuHelp.setObjectName(_fromUtf8("menuHelp"))
-        self.menuClassification = QtGui.QMenu(self.menuBar)
-        self.menuClassification.setObjectName(_fromUtf8("menuClassification"))
-        self.menuSupervised = QtGui.QMenu(self.menuClassification)
-        self.menuSupervised.setObjectName(_fromUtf8("menuSupervised"))
-        self.menuClustering = QtGui.QMenu(self.menuClassification)
-        self.menuClustering.setObjectName(_fromUtf8("menuClustering"))
-        self.menuVisualization = QtGui.QMenu(self.menuBar)
-        self.menuVisualization.setObjectName(_fromUtf8("menuVisualization"))
-        MainWindow.setMenuBar(self.menuBar)
-        self.actionLoad_reference_Data = QtGui.QAction(MainWindow)
-        self.actionLoad_reference_Data.setObjectName(_fromUtf8("actionLoad_reference_Data"))
-        self.actionLoad_Unknown_Data = QtGui.QAction(MainWindow)
-        self.actionLoad_Unknown_Data.setObjectName(_fromUtf8("actionLoad_Unknown_Data"))
-        self.actionSave_Current_Workflow = QtGui.QAction(MainWindow)
-        self.actionSave_Current_Workflow.setObjectName(_fromUtf8("actionSave_Current_Workflow"))
-        self.actionSave_Current_Plots = QtGui.QAction(MainWindow)
-        self.actionSave_Current_Plots.setObjectName(_fromUtf8("actionSave_Current_Plots"))
-        self.actionSave_Current_Data = QtGui.QAction(MainWindow)
-        self.actionSave_Current_Data.setObjectName(_fromUtf8("actionSave_Current_Data"))
-        self.actionCreate_New_Workflow = QtGui.QAction(MainWindow)
-        self.actionCreate_New_Workflow.setObjectName(_fromUtf8("actionCreate_New_Workflow"))
-        self.actionNoise_Reduction = QtGui.QAction(MainWindow)
-        self.actionNoise_Reduction.setObjectName(_fromUtf8("actionNoise_Reduction"))
-        self.actionApply_Mask = QtGui.QAction(MainWindow)
-        self.actionApply_Mask.setObjectName(_fromUtf8("actionApply_Mask"))
-        self.actionInterpolate = QtGui.QAction(MainWindow)
-        self.actionInterpolate.setObjectName(_fromUtf8("actionInterpolate"))
-        self.actionInstrument_Response = QtGui.QAction(MainWindow)
-        self.actionInstrument_Response.setObjectName(_fromUtf8("actionInstrument_Response"))
-        self.actionALS = QtGui.QAction(MainWindow)
-        self.actionALS.setObjectName(_fromUtf8("actionALS"))
-        self.actionDietrich = QtGui.QAction(MainWindow)
-        self.actionDietrich.setObjectName(_fromUtf8("actionDietrich"))
-        self.actionPolyFit = QtGui.QAction(MainWindow)
-        self.actionPolyFit.setObjectName(_fromUtf8("actionPolyFit"))
-        self.actionAirPLS = QtGui.QAction(MainWindow)
-        self.actionAirPLS.setObjectName(_fromUtf8("actionAirPLS"))
-        self.actionFABC = QtGui.QAction(MainWindow)
-        self.actionFABC.setObjectName(_fromUtf8("actionFABC"))
-        self.actionKK = QtGui.QAction(MainWindow)
-        self.actionKK.setObjectName(_fromUtf8("actionKK"))
-        self.actionMario = QtGui.QAction(MainWindow)
-        self.actionMario.setObjectName(_fromUtf8("actionMario"))
-        self.actionMedian = QtGui.QAction(MainWindow)
-        self.actionMedian.setObjectName(_fromUtf8("actionMedian"))
-        self.actionRubberband = QtGui.QAction(MainWindow)
-        self.actionRubberband.setObjectName(_fromUtf8("actionRubberband"))
-        self.actionUndecimated_Wavelet = QtGui.QAction(MainWindow)
-        self.actionUndecimated_Wavelet.setObjectName(_fromUtf8("actionUndecimated_Wavelet"))
-        self.actionRatio = QtGui.QAction(MainWindow)
-        self.actionRatio.setObjectName(_fromUtf8("actionRatio"))
-        self.actionTommy_s_Methgod = QtGui.QAction(MainWindow)
-        self.actionTommy_s_Methgod.setObjectName(_fromUtf8("actionTommy_s_Methgod"))
-        self.actionPiecewise_Direct_Standardization = QtGui.QAction(MainWindow)
-        self.actionPiecewise_Direct_Standardization.setObjectName(
-            _fromUtf8("actionPiecewise_Direct_Standardization"))
-        self.actionPCA = QtGui.QAction(MainWindow)
-        self.actionPCA.setObjectName(_fromUtf8("actionPCA"))
-        self.actionICA = QtGui.QAction(MainWindow)
-        self.actionICA.setObjectName(_fromUtf8("actionICA"))
-        self.actionK_Means = QtGui.QAction(MainWindow)
-        self.actionK_Means.setObjectName(_fromUtf8("actionK_Means"))
-        self.actionHierarchical = QtGui.QAction(MainWindow)
-        self.actionHierarchical.setObjectName(_fromUtf8("actionHierarchical"))
-        self.actionOthers = QtGui.QAction(MainWindow)
-        self.actionOthers.setObjectName(_fromUtf8("actionOthers"))
-        self.actionOthers_2 = QtGui.QAction(MainWindow)
-        self.actionOthers_2.setObjectName(_fromUtf8("actionOthers_2"))
-        self.actionOthers_3 = QtGui.QAction(MainWindow)
-        self.actionOthers_3.setObjectName(_fromUtf8("actionOthers_3"))
-        self.actionPLS = QtGui.QAction(MainWindow)
-        self.actionPLS.setObjectName(_fromUtf8("actionPLS"))
-        self.actionSM_PLS = QtGui.QAction(MainWindow)
-        self.actionSM_PLS.setObjectName(_fromUtf8("actionSM_PLS"))
-        self.actionICA_Regression = QtGui.QAction(MainWindow)
-        self.actionICA_Regression.setObjectName(_fromUtf8("actionICA_Regression"))
-        self.actionGaussian_Process = QtGui.QAction(MainWindow)
-        self.actionGaussian_Process.setObjectName(_fromUtf8("actionGaussian_Process"))
-        self.actionMLP = QtGui.QAction(MainWindow)
-        self.actionMLP.setObjectName(_fromUtf8("actionMLP"))
-        self.actionSVM = QtGui.QAction(MainWindow)
-        self.actionSVM.setObjectName(_fromUtf8("actionSVM"))
-        self.actionOthers_4 = QtGui.QAction(MainWindow)
-        self.actionOthers_4.setObjectName(_fromUtf8("actionOthers_4"))
-        self.actionOthers_5 = QtGui.QAction(MainWindow)
-        self.actionOthers_5.setObjectName(_fromUtf8("actionOthers_5"))
-        self.actionIndex = QtGui.QAction(MainWindow)
-        self.actionIndex.setObjectName(_fromUtf8("actionIndex"))
-        self.actionContent_2 = QtGui.QAction(MainWindow)
-        self.actionContent_2.setObjectName(_fromUtf8("actionContent_2"))
-        self.actionAbout = QtGui.QAction(MainWindow)
-        self.actionAbout.setObjectName(_fromUtf8("actionAbout"))
-        self.actionAbout_QtCreator = QtGui.QAction(MainWindow)
-        self.actionAbout_QtCreator.setObjectName(_fromUtf8("actionAbout_QtCreator"))
-        self.actionExit = QtGui.QAction(MainWindow)
-        self.actionExit.setObjectName(_fromUtf8("actionExit"))
-        self.actionNormalization = QtGui.QAction(MainWindow)
-        self.actionNormalization.setObjectName(_fromUtf8("actionNormalization"))
-        self.actionICA_2 = QtGui.QAction(MainWindow)
-        self.actionICA_2.setObjectName(_fromUtf8("actionICA_2"))
-        self.actionPCA_2 = QtGui.QAction(MainWindow)
-        self.actionPCA_2.setObjectName(_fromUtf8("actionPCA_2"))
-        self.actionPLS_DA = QtGui.QAction(MainWindow)
-        self.actionPLS_DA.setObjectName(_fromUtf8("actionPLS_DA"))
-        self.actionSIMCA = QtGui.QAction(MainWindow)
-        self.actionSIMCA.setObjectName(_fromUtf8("actionSIMCA"))
-        self.actionK_means = QtGui.QAction(MainWindow)
-        self.actionK_means.setObjectName(_fromUtf8("actionK_means"))
-        self.actionHierarchical_2 = QtGui.QAction(MainWindow)
-        self.actionHierarchical_2.setObjectName(_fromUtf8("actionHierarchical_2"))
-        self.actionCross_Validation = QtGui.QAction(MainWindow)
-        self.actionCross_Validation.setObjectName(_fromUtf8("actionCross_Validation"))
-        self.actionTrain = QtGui.QAction(MainWindow)
-        self.actionTrain.setObjectName(_fromUtf8("actionTrain"))
-        self.actionPredict = QtGui.QAction(MainWindow)
-        self.actionPredict.setObjectName(_fromUtf8("actionPredict"))
-        self.actionLine_Plot = QtGui.QAction(MainWindow)
-        self.actionLine_Plot.setObjectName(_fromUtf8("actionLine_Plot"))
-        self.action1_to_1_Plot = QtGui.QAction(MainWindow)
-        self.action1_to_1_Plot.setObjectName(_fromUtf8("action1_to_1_Plot"))
-        self.actionScatter_Plot = QtGui.QAction(MainWindow)
-        self.actionScatter_Plot.setObjectName(_fromUtf8("actionScatter_Plot"))
-        self.actionSet_output_location = QtGui.QAction(MainWindow)
-        self.actionSet_output_location.setObjectName(_fromUtf8("actionSet_output_location"))
-        self.actionCreate_N_Folds = QtGui.QAction(MainWindow)
-        self.actionCreate_N_Folds.setObjectName(_fromUtf8("actionCreate_N_Folds"))
-        self.actionCreate_Test_Folds = QtGui.QAction(MainWindow)
-        self.actionCreate_Test_Folds.setObjectName(_fromUtf8("actionCreate_Test_Folds"))
-        self.actionStratified_Folds = QtGui.QAction(MainWindow)
-        self.actionStratified_Folds.setObjectName(_fromUtf8("actionStratified_Folds"))
-        # self.actionTrain_Submodels = QtGui.QAction(MainWindow)
-        # self.actionTrain_Submodels.setObjectName(_fromUtf8("actionTrain_Submodels"))
-        self.menuFile.addAction(self.actionLoad_reference_Data)
-        self.menuFile.addAction(self.actionLoad_Unknown_Data)
-        self.menuFile.addAction(self.actionSet_output_location)
-        self.menuFile.addSeparator()
-        self.menuFile.addAction(self.actionSave_Current_Plots)
-        self.menuFile.addAction(self.actionSave_Current_Data)
-        self.menuFile.addSeparator()
-        self.menuFile.addAction(self.actionCreate_New_Workflow)
-        self.menuFile.addAction(self.actionSave_Current_Workflow)
-        self.menuFile.addSeparator()
-        self.menuFile.addAction(self.actionExit)
-        self.menuBaseline_Removal.addAction(self.actionALS)
-        self.menuBaseline_Removal.addAction(self.actionDietrich)
-        self.menuBaseline_Removal.addAction(self.actionPolyFit)
-        self.menuBaseline_Removal.addAction(self.actionAirPLS)
-        self.menuBaseline_Removal.addAction(self.actionFABC)
-        self.menuBaseline_Removal.addAction(self.actionKK)
-        self.menuBaseline_Removal.addAction(self.actionMario)
-        self.menuBaseline_Removal.addAction(self.actionMedian)
-        self.menuBaseline_Removal.addAction(self.actionRubberband)
-        self.menuBaseline_Removal.addAction(self.actionUndecimated_Wavelet)
-        self.menuCalibration_Transfer.addAction(self.actionRatio)
-        self.menuCalibration_Transfer.addAction(self.actionTommy_s_Methgod)
-        self.menuCalibration_Transfer.addAction(self.actionPiecewise_Direct_Standardization)
-        self.menuCalibration_Transfer.addAction(self.actionOthers_3)
-        self.menuPreprocessing.addAction(self.actionNormalization)
-        self.menuPreprocessing.addAction(self.actionNoise_Reduction)
-        self.menuPreprocessing.addAction(self.actionApply_Mask)
-        self.menuPreprocessing.addAction(self.actionInterpolate)
-        self.menuPreprocessing.addAction(self.actionInstrument_Response)
-        self.menuPreprocessing.addAction(self.menuBaseline_Removal.menuAction())
-        self.menuPreprocessing.addAction(self.menuCalibration_Transfer.menuAction())
-        self.menuPreprocessing.addAction(self.actionICA_2)
-        self.menuPreprocessing.addAction(self.actionPCA_2)
-        self.menuPreprocessing.addAction(self.actionStratified_Folds)
-        self.menuRegression.addAction(self.actionCross_Validation)
-        self.menuRegression.addAction(self.actionTrain)
-        # self.menuRegression.addAction(self.actionTrain_Submodels)
-        self.menuRegression.addAction(self.actionPredict)
-        self.menuHelp.addAction(self.actionIndex)
-        self.menuHelp.addAction(self.actionContent_2)
-        self.menuHelp.addSeparator()
-        self.menuHelp.addAction(self.actionAbout)
-        self.menuHelp.addAction(self.actionAbout_QtCreator)
-        self.menuSupervised.addAction(self.actionPLS_DA)
-        self.menuSupervised.addAction(self.actionSIMCA)
-        self.menuClustering.addAction(self.actionK_means)
-        self.menuClustering.addAction(self.actionHierarchical_2)
-        self.menuClassification.addAction(self.menuSupervised.menuAction())
-        self.menuClassification.addAction(self.menuClustering.menuAction())
-        self.menuVisualization.addAction(self.actionLine_Plot)
-        self.menuVisualization.addAction(self.action1_to_1_Plot)
-        self.menuVisualization.addAction(self.actionScatter_Plot)
-        self.menuBar.addAction(self.menuFile.menuAction())
-        self.menuBar.addAction(self.menuPreprocessing.menuAction())
-        self.menuBar.addAction(self.menuClassification.menuAction())
-        self.menuBar.addAction(self.menuRegression.menuAction())
-        self.menuBar.addAction(self.menuVisualization.menuAction())
-        self.menuBar.addAction(self.menuHelp.menuAction())
-        QtCore.QMetaObject.connectSlotsByName(MainWindow)
+        mainframe_.mainframe(self, MainWindow)
 
-        MainWindow.setWindowTitle(_translate("MainWindow", "PYSAT", None))
-        self.menuFile.setTitle(_translate("MainWindow", "File", None))
-        self.menuPreprocessing.setTitle(_translate("MainWindow", "Preprocessing", None))
-        self.menuBaseline_Removal.setTitle(_translate("MainWindow", "Baseline Removal", None))
-        self.menuCalibration_Transfer.setTitle(_translate("MainWindow", "Calibration Transfer", None))
-        self.menuRegression.setTitle(_translate("MainWindow", "Regression", None))
-        self.menuHelp.setTitle(_translate("MainWindow", "Help", None))
-        self.menuClassification.setTitle(_translate("MainWindow", "Classification", None))
-        self.menuSupervised.setTitle(_translate("MainWindow", "Supervised", None))
-        self.menuClustering.setTitle(_translate("MainWindow", "Clustering", None))
-        self.menuVisualization.setTitle(_translate("MainWindow", "Visualization", None))
-        self.actionLoad_reference_Data.setText(_translate("MainWindow", "Load Reference Data", None))
-        self.actionLoad_Unknown_Data.setText(_translate("MainWindow", "Load Unknown Data", None))
-        self.actionSave_Current_Workflow.setText(_translate("MainWindow", "Save Current Workflow", None))
-        self.actionSave_Current_Plots.setText(_translate("MainWindow", "Save Current Plots", None))
-        self.actionSave_Current_Data.setText(_translate("MainWindow", "Save Current Data", None))
-        self.actionCreate_New_Workflow.setText(_translate("MainWindow", "Create New Workflow", None))
-        self.actionNoise_Reduction.setText(_translate("MainWindow", "Noise Reduction", None))
-        self.actionApply_Mask.setText(_translate("MainWindow", "Apply Mask", None))
-        self.actionInterpolate.setText(_translate("MainWindow", "Interpolate (unknown to known)", None))
-        self.actionInstrument_Response.setText(_translate("MainWindow", "Instrument Response", None))
-        self.actionALS.setText(_translate("MainWindow", "ALS", None))
-        self.actionDietrich.setText(_translate("MainWindow", "Dietrich", None))
-        self.actionPolyFit.setText(_translate("MainWindow", "PolyFit", None))
-        self.actionAirPLS.setText(_translate("MainWindow", "AirPLS", None))
-        self.actionFABC.setText(_translate("MainWindow", "FABC", None))
-        self.actionKK.setText(_translate("MainWindow", "KK", None))
-        self.actionMario.setText(_translate("MainWindow", "Mario", None))
-        self.actionMedian.setText(_translate("MainWindow", "Median", None))
-        self.actionRubberband.setText(_translate("MainWindow", "Rubberband", None))
-        self.actionUndecimated_Wavelet.setText(_translate("MainWindow", "Undecimated Wavelet", None))
-        self.actionRatio.setText(_translate("MainWindow", "Ratio", None))
-        self.actionTommy_s_Methgod.setText(_translate("MainWindow", "Tommy\'s Method", None))
-        self.actionPiecewise_Direct_Standardization.setText(
-            _translate("MainWindow", "Piecewise Direct Standardization", None))
-        self.actionPCA.setText(_translate("MainWindow", "PCA", None))
-        self.actionICA.setText(_translate("MainWindow", "ICA", None))
-        self.actionK_Means.setText(_translate("MainWindow", "K-Means", None))
-        self.actionHierarchical.setText(_translate("MainWindow", "Hierarchical", None))
-        self.actionOthers.setText(_translate("MainWindow", "Others...", None))
-        self.actionOthers_2.setText(_translate("MainWindow", "Others...", None))
-        self.actionOthers_3.setText(_translate("MainWindow", "Others...", None))
-        self.actionPLS.setText(_translate("MainWindow", "PLS", None))
-        self.actionSM_PLS.setText(_translate("MainWindow", "SM-PLS", None))
-        self.actionICA_Regression.setText(_translate("MainWindow", "ICA Regression", None))
-        self.actionGaussian_Process.setText(_translate("MainWindow", "Gaussian Process", None))
-        self.actionMLP.setText(_translate("MainWindow", "MLP", None))
-        self.actionSVM.setText(_translate("MainWindow", "SVM", None))
-        self.actionOthers_4.setText(_translate("MainWindow", "Others...", None))
-        self.actionOthers_5.setText(_translate("MainWindow", "Others...", None))
-        self.actionIndex.setText(_translate("MainWindow", "Index", None))
-        self.actionContent_2.setText(_translate("MainWindow", "Content", None))
-        self.actionAbout.setText(_translate("MainWindow", "About...", None))
-        self.actionAbout_QtCreator.setText(_translate("MainWindow", "About Qt...", None))
-        self.actionExit.setText(_translate("MainWindow", "Exit", None))
-        self.actionNormalization.setText(_translate("MainWindow", "Normalization", None))
-        self.actionICA_2.setText(_translate("MainWindow", "ICA", None))
-        self.actionPCA_2.setText(_translate("MainWindow", "PCA", None))
-        self.actionPLS_DA.setText(_translate("MainWindow", "PLS-DA", None))
-        self.actionSIMCA.setText(_translate("MainWindow", "SIMCA", None))
-        self.actionK_means.setText(_translate("MainWindow", "K-means", None))
-        self.actionHierarchical_2.setText(_translate("MainWindow", "Hierarchical", None))
-        self.actionCross_Validation.setText(_translate("MainWindow", "Cross Validation", None))
-        self.actionTrain.setText(_translate("MainWindow", "Train", None))
-        self.actionPredict.setText(_translate("MainWindow", "Predict", None))
-        self.actionLine_Plot.setText(_translate("MainWindow", "Line Plot", None))
-        self.action1_to_1_Plot.setText(_translate("MainWindow", "1 to 1 Plot", None))
-        self.actionScatter_Plot.setText(_translate("MainWindow", "Scatter Plot", None))
-        self.actionSet_output_location.setText(_translate("MainWindow", "Output Location", None))
-        self.actionCreate_N_Folds.setText(_translate("MainWindow", "Create N Folds", None))
-        self.actionStratified_Folds.setText(_translate("MainWindow", "Stratified Folds", None))
+    def file_outpath(self, MainWindow):
+        file_outpath_.file_outpath(self, MainWindow)
 
-    def unknown_data(self, MainWindow):
-        self.unknownData = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.unknownData.setFont(font)
-        self.unknownData.setObjectName(_fromUtf8("unknownData"))
-        self.verticalLayout_6 = QtGui.QVBoxLayout(self.unknownData)
-        self.verticalLayout_6.setMargin(11)
-        self.verticalLayout_6.setSpacing(6)
-        self.verticalLayout_6.setObjectName(_fromUtf8("verticalLayout_6"))
-        self.horizontalLayout = QtGui.QHBoxLayout()
-        self.horizontalLayout.setMargin(11)
-        self.horizontalLayout.setSpacing(6)
-        self.horizontalLayout.setObjectName(_fromUtf8("horizontalLayout"))
-        self.verticalLayout = QtGui.QVBoxLayout()
-        self.verticalLayout.setMargin(11)
-        self.verticalLayout.setSpacing(6)
-        self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
-        self.label = QtGui.QLabel(self.unknownData)
-        self.verticalLayout.addWidget(self.label)
-        self.horizontalLayout.addLayout(self.verticalLayout)
-        self.verticalLayout_3 = QtGui.QVBoxLayout()
-        self.verticalLayout_3.setMargin(11)
-        self.verticalLayout_3.setSpacing(6)
-        self.verticalLayout_3.setObjectName(_fromUtf8("verticalLayout_3"))
-        self.lineEdit_2 = QtGui.QLineEdit(self.unknownData)
-        self.lineEdit_2.setReadOnly(True)
-        self.lineEdit_2.setObjectName(_fromUtf8("lineEdit_2"))
-        self.verticalLayout_3.addWidget(self.lineEdit_2)
-        self.horizontalLayout.addLayout(self.verticalLayout_3)
-        self.verticalLayout_5 = QtGui.QVBoxLayout()
-        self.verticalLayout_5.setMargin(11)
-        self.verticalLayout_5.setSpacing(6)
-        self.verticalLayout_5.setObjectName(_fromUtf8("verticalLayout_5"))
-        self.unknownDataButton = QtGui.QToolButton(self.unknownData)
-        self.unknownDataButton.setObjectName(_fromUtf8("unknownDataButton"))
-        self.verticalLayout_5.addWidget(self.unknownDataButton)
-        self.horizontalLayout.addLayout(self.verticalLayout_5)
-        self.verticalLayout_6.addLayout(self.horizontalLayout)
-        self.verticalLayout_8.addWidget(self.unknownData)
-        self.unknownData.raise_()
-        self.unknownData.setTitle(_translate("MainWindow", "File", None))
-        self.label.setText(_translate("MainWindow", "Unknown Data", None))
-        self.lineEdit_2.setText(_translate("MainWindow", "*.csv", None))
-        self.unknownDataButton.setText(_translate("MainWindow", "...", None))
-        try:
-            self.unknownDataButton.clicked.connect(
-                lambda: pysat_ui.on_getDataButton_clicked(self, self.lineEdit_2, 'Unknown Data'))
-        except:
-            pass
+    def get_unknown_data(self, MainWindow):
+        get_data_.get_data_u(self, MainWindow)
 
-    def known_data(self, MainWindow):
-        self.referenceData = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.referenceData.setFont(font)
-        self.referenceData.setObjectName(_fromUtf8("referenceData"))
-        self.verticalLayout_15 = QtGui.QVBoxLayout(self.referenceData)
-        self.verticalLayout_15.setMargin(11)
-        self.verticalLayout_15.setSpacing(6)
-        self.verticalLayout_15.setObjectName(_fromUtf8("verticalLayout_15"))
-        self.horizontalLayout_11 = QtGui.QHBoxLayout()
-        self.horizontalLayout_11.setMargin(11)
-        self.horizontalLayout_11.setSpacing(6)
-        self.horizontalLayout_11.setObjectName(_fromUtf8("horizontalLayout_11"))
-        self.verticalLayout_16 = QtGui.QVBoxLayout()
-        self.verticalLayout_16.setMargin(11)
-        self.verticalLayout_16.setSpacing(6)
-        self.verticalLayout_16.setObjectName(_fromUtf8("verticalLayout_16"))
-        self.label_8 = QtGui.QLabel(self.referenceData)
-        self.label_8.setObjectName(_fromUtf8("label_8"))
-        self.verticalLayout_16.addWidget(self.label_8)
-        self.horizontalLayout_11.addLayout(self.verticalLayout_16)
-        self.verticalLayout_20 = QtGui.QVBoxLayout()
-        self.verticalLayout_20.setMargin(11)
-        self.verticalLayout_20.setSpacing(6)
-        self.verticalLayout_20.setObjectName(_fromUtf8("verticalLayout_20"))
-        self.lineEdit_9 = QtGui.QLineEdit(self.referenceData)
-        self.lineEdit_9.setReadOnly(True)
-        self.lineEdit_9.setObjectName(_fromUtf8("lineEdit_9"))
-        self.verticalLayout_20.addWidget(self.lineEdit_9)
-        self.horizontalLayout_11.addLayout(self.verticalLayout_20)
-        self.verticalLayout_21 = QtGui.QVBoxLayout()
-        self.verticalLayout_21.setMargin(11)
-        self.verticalLayout_21.setSpacing(6)
-        self.verticalLayout_21.setObjectName(_fromUtf8("verticalLayout_21"))
-        self.fullDataBaseButton_3 = QtGui.QToolButton(self.referenceData)
-        self.fullDataBaseButton_3.setObjectName(_fromUtf8("fullDataBaseButton_3"))
-        self.verticalLayout_21.addWidget(self.fullDataBaseButton_3)
-        self.horizontalLayout_11.addLayout(self.verticalLayout_21)
-        self.verticalLayout_15.addLayout(self.horizontalLayout_11)
-        self.verticalLayout_8.addWidget(self.referenceData)
-        self.referenceData.raise_()
-        self.referenceData.setTitle(_translate("MainWindow", "File", None))
-        self.label_8.setText(_translate("MainWindow", "Full Database", None))
-        self.lineEdit_9.setText(_translate("MainWindow", "*.csv", None))
-        self.fullDataBaseButton_3.setText(_translate("MainWindow", "...", None))
-        try:
-            self.fullDataBaseButton_3.clicked.connect(
-                lambda: pysat_ui.on_getDataButton_clicked(self, self.lineEdit_9, 'Known Data'))
-        except:
-            pass
+    def get_known_data(self, MainWindow):
+        get_data_.get_data_u(self, MainWindow)
 
-    def output_folder(self, MainWindow):
-        self.outputFolder = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.outputFolder.setFont(font)
-        self.outputFolder.setObjectName(_fromUtf8("outputFolder"))
-        self.verticalLayout_22 = QtGui.QVBoxLayout(self.outputFolder)
-        self.verticalLayout_22.setMargin(11)
-        self.verticalLayout_22.setSpacing(6)
-        self.verticalLayout_22.setObjectName(_fromUtf8("verticalLayout_22"))
-        self.horizontalLayout_13 = QtGui.QHBoxLayout()
-        self.horizontalLayout_13.setMargin(11)
-        self.horizontalLayout_13.setSpacing(6)
-        self.horizontalLayout_13.setObjectName(_fromUtf8("horizontalLayout_13"))
-        self.verticalLayout_25 = QtGui.QVBoxLayout()
-        self.verticalLayout_25.setMargin(11)
-        self.verticalLayout_25.setSpacing(6)
-        self.verticalLayout_25.setObjectName(_fromUtf8("verticalLayout_25"))
-        self.label_9 = QtGui.QLabel(self.outputFolder)
-        self.label_9.setObjectName(_fromUtf8("label_9"))
-        self.verticalLayout_25.addWidget(self.label_9)
-        self.horizontalLayout_13.addLayout(self.verticalLayout_25)
-        self.verticalLayout_28 = QtGui.QVBoxLayout()
-        self.verticalLayout_28.setMargin(11)
-        self.verticalLayout_28.setSpacing(6)
-        self.verticalLayout_28.setObjectName(_fromUtf8("verticalLayout_28"))
-        self.lineEdit_10 = QtGui.QLineEdit(self.outputFolder)
-        self.lineEdit_10.setReadOnly(True)
-        self.lineEdit_10.setObjectName(_fromUtf8("lineEdit_10"))
-        self.verticalLayout_28.addWidget(self.lineEdit_10)
-        self.horizontalLayout_13.addLayout(self.verticalLayout_28)
-        self.verticalLayout_29 = QtGui.QVBoxLayout()
-        self.verticalLayout_29.setMargin(11)
-        self.verticalLayout_29.setSpacing(6)
-        self.verticalLayout_29.setObjectName(_fromUtf8("verticalLayout_29"))
-        self.fullDataBaseButton_4 = QtGui.QToolButton(self.outputFolder)
-        self.fullDataBaseButton_4.setObjectName(_fromUtf8("fullDataBaseButton_4"))
-        self.verticalLayout_29.addWidget(self.fullDataBaseButton_4)
-        self.horizontalLayout_13.addLayout(self.verticalLayout_29)
-        self.verticalLayout_22.addLayout(self.horizontalLayout_13)
-        self.verticalLayout_8.addWidget(self.outputFolder)
-        self.outputFolder.raise_()
-        self.outputFolder.setTitle(_translate("MainWindow", "Output", None))
-        self.label_9.setText(_translate("MainWindow", "Output Folder", None))
-        self.lineEdit_10.setText(_translate("MainWindow", "*.*", None))
-        self.fullDataBaseButton_4.setText(_translate("MainWindow", "...", None))
-        try:
-            self.fullDataBaseButton_4.clicked.connect(
-                lambda: pysat_ui.on_outPutLocationButton_clicked(self, self.lineEdit_10))
-        except:
-            pass
-
-            #### Opening Files
-
-    def regression_train(self, MainWindow):
-        self.regression_train = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.regression_train.setFont(font)
-        self.regression_train.setObjectName(_fromUtf8("regression_train"))
-        self.regression_vlayout = QtGui.QVBoxLayout(self.regression_train)
-        self.regression_vlayout.setObjectName(_fromUtf8("regression_vlayout"))
-        # choose data
-        self.regression_choosedata_hlayout = QtGui.QHBoxLayout()
-        self.regression_choosedata_hlayout.setObjectName(_fromUtf8("regression_choosedata_hlayout"))
-        self.regression_train_choosedata_label = QtGui.QLabel(self.regression_train)
-        self.regression_train_choosedata_label.setObjectName(_fromUtf8("regression_train_choosedata_label"))
-        self.regression_train_choosedata_label.setText(_translate("regression_train", "Choose data:", None))
-        self.regression_choosedata_hlayout.addWidget(self.regression_train_choosedata_label)
-        datachoices = self.pysat_fun.datakeys
-        if datachoices == []:
-            datachoices = ['No data has been loaded!']
-        self.regression_choosedata = make_combobox(datachoices)
-        self.regression_choosedata.setIconSize(QtCore.QSize(50, 20))
-        self.regression_choosedata.setObjectName(_fromUtf8("regression_choosedata"))
-        self.regression_choosedata_hlayout.addWidget(self.regression_choosedata)
-        spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.regression_choosedata_hlayout.addItem(spacerItem)
-        self.regression_vlayout.addLayout(self.regression_choosedata_hlayout)
-        # choose variables
-        self.regression_choosevars_hlayout = QtGui.QHBoxLayout()
-        self.regression_choosevars_hlayout.setObjectName(_fromUtf8("regression_choosevars_hlayout"))
-        self.regression_train_choosex_label = QtGui.QLabel(self.regression_train)
-        self.regression_train_choosex_label.setObjectName(_fromUtf8("regression_train_choosex_label"))
-        self.regression_choosevars_hlayout.addWidget(self.regression_train_choosex_label)
-
-        xvarchoices = self.pysat_fun.data[self.regression_choosedata.currentText()].df.columns.levels[0].values
-        self.regression_train_choosex = make_listwidget(xvarchoices)
-        self.regression_train_choosex.setObjectName(_fromUtf8("regression_train_choosex"))
-        self.regression_choosevars_hlayout.addWidget(self.regression_train_choosex)
-        self.regression_train_choosey_label = QtGui.QLabel(self.regression_train)
-        self.regression_train_choosey_label.setObjectName(_fromUtf8("regression_train_choosey_label"))
-        self.regression_choosevars_hlayout.addWidget(self.regression_train_choosey_label)
-
-        yvarchoices = self.pysat_fun.data[self.regression_choosedata.currentText()].df['comp'].columns.values
-        self.regression_train_choosey = make_listwidget(yvarchoices)
-        self.regression_choosevars_hlayout.addWidget(self.regression_train_choosey)
-        spacerItem1 = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.regression_choosevars_hlayout.addItem(spacerItem1)
-        self.regression_vlayout.addLayout(self.regression_choosevars_hlayout)
-
-        # ransac options
-        self.ransac_hlayout = QtGui.QHBoxLayout()
-        self.regression_ransac_checkbox = QtGui.QCheckBox(self.regression_train)
-        self.regression_ransac_checkbox.setObjectName(_fromUtf8("regression_ransac_checkbox"))
-        self.regression_ransac_checkbox.setText('RANSAC')
-        self.ransac_hlayout.addWidget(self.regression_ransac_checkbox)
-        self.regression_vlayout.addLayout(self.ransac_hlayout)
-
-        self.regression_ransac_checkbox.stateChanged.connect(self.make_ransac_widget)
-        # choose regression algorithm
-        self.regression_choosealg_hlayout = QtGui.QHBoxLayout()
-        self.regression_choosealg_hlayout.setObjectName(_fromUtf8("regression_choosealg_hlayout"))
-        self.regression_choosealg_label = QtGui.QLabel(self.regression_train)
-        self.regression_choosealg_label.setObjectName(_fromUtf8("regression_choosealg_label"))
-        self.regression_choosealg_hlayout.addWidget(self.regression_choosealg_label)
-        self.regression_alg_choices = ['Choose an algorithm', 'PLS', 'GP', 'More to come...']
-        self.regression_choosealg = make_combobox(self.regression_alg_choices)
-        self.regression_choosealg.setIconSize(QtCore.QSize(50, 20))
-        self.regression_choosealg.setObjectName(_fromUtf8("regression_choosealg"))
-        self.regression_choosealg_hlayout.addWidget(self.regression_choosealg)
-        regression_choosealg_spacer = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.regression_choosealg_hlayout.addItem(regression_choosealg_spacer)
-        self.regression_vlayout.addLayout(self.regression_choosealg_hlayout)
-        self.regression_choosealg.activated.connect(self.make_regression_widget)
-
-        self.verticalLayout_8.addWidget(self.regression_train)
-        self.regression_train.raise_()
-        self.regression_train.setTitle(_translate("regression_train", "Regression - Train", None))
-
-    def stratified_folds(self, MainWindow):
-        self.strat_folds = QtGui.QGroupBox(self.scrollAreaWidgetContents_2)
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.strat_folds.setFont(font)
-        self.strat_folds.setObjectName(_fromUtf8("Stratified Folds"))
-        self.strat_folds_vlayout = QtGui.QVBoxLayout(self.strat_folds)
-        self.strat_folds_vlayout.setObjectName(_fromUtf8("strat_folds_vlayout"))
-        self.strat_folds_choose_data_label = QtGui.QLabel(self.strat_folds)
-        self.strat_folds_choose_data_label.setObjectName(_fromUtf8("strat_folds_choose_data_label"))
-        self.strat_folds_vlayout.addWidget(self.strat_folds_choose_data_label)
-        datachoices = self.pysat_fun.datakeys
-        if datachoices == []:
-            datachoices = ['No data has been loaded!']
-        self.strat_folds_choose_data = make_combobox(datachoices)
-        self.strat_folds_vlayout.addWidget(self.strat_folds_choose_data)
-        self.strat_folds_choose_var_label = QtGui.QLabel(self.strat_folds)
-        self.strat_folds_choose_var_label.setObjectName(_fromUtf8("strat_folds_choose_var_label"))
-        self.strat_folds_vlayout.addWidget(self.strat_folds_choose_var_label)
-
-        varchoices = self.pysat_fun.data[self.strat_folds_choose_data.currentText()].df['meta'].columns.values
-        self.strat_folds_choose_var = make_combobox(varchoices)
-        self.strat_folds_vlayout.addWidget(self.strat_folds_choose_var)
-        self.strat_folds_choose_data.activated[int].connect(self.strat_fold_change_vars)
-        self.strat_folds_hlayout = QtGui.QHBoxLayout()
-        self.strat_folds_hlayout.setObjectName(_fromUtf8("strat_folds_hlayout"))
-        self.nfolds_label = QtGui.QLabel(self.strat_folds)
-        self.nfolds_label.setObjectName(_fromUtf8("nfolds_label"))
-        self.strat_folds_hlayout.addWidget(self.nfolds_label)
-        self.nfolds_spin = QtGui.QSpinBox(self.strat_folds)
-        self.nfolds_spin.setObjectName(_fromUtf8("nfolds_spin"))
-        self.nfolds_spin.setMinimum(1)
-        self.strat_folds_hlayout.addWidget(self.nfolds_spin)
-        self.test_fold_label = QtGui.QLabel(self.strat_folds)
-        self.test_fold_label.setObjectName(_fromUtf8("test_fold_label"))
-        self.strat_folds_hlayout.addWidget(self.test_fold_label)
-        foldchoices = ['1']
-        self.choose_test_fold = make_combobox(foldchoices)
-        self.choose_test_fold.setObjectName(_fromUtf8("choose_test_fold"))
-        self.nfolds_spin.valueChanged.connect(self.strat_fold_change_testfolds)
-        self.strat_folds_hlayout.addWidget(self.choose_test_fold)
-        self.spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.strat_folds_hlayout.addItem(self.spacerItem)
-        self.create_folds = QtGui.QPushButton(self.strat_folds)
-        self.create_folds.setObjectName(_fromUtf8("create_folds"))
-        self.create_folds.setText(_translate("strat_folds", "Create Folds", None))
-        self.strat_folds_hlayout.addWidget(self.create_folds)
-        self.strat_folds_vlayout.addLayout(self.strat_folds_hlayout)
-        self.verticalLayout_8.addWidget(self.strat_folds)
-        self.strat_folds.raise_()
-        self.strat_folds.setTitle(_translate("MainWindow", "Stratified Folds", None))
-        self.nfolds_label.setText(_translate("strat_folds", "N folds", None))
-        self.test_fold_label.setText(_translate("strat_folds", "Test Fold", None))
-        self.create_folds.setText(_translate("strat_folds", "Create Folds", None))
-        self.strat_folds_choose_data_label.setText(_translate("strat_folds", "Choose data to stratify:", None))
-        self.strat_folds_choose_var_label.setText(_translate("strat_folds", "Choose variable on which to sort:", None))
-        try:
-            self.create_folds.clicked.connect(
-                lambda: self.pysat_fun.do_strat_folds(datakey=str(self.strat_folds_choose_data.currentText()),
-                                                      nfolds=int(self.nfolds_spin.text()),
-                                                      testfold=int(self.choose_test_fold.currentText()),
-                                                      colname=('comp', self.strat_folds_choose_var.currentText())))
-        except:
-            print('There was a problem with creating stratified folds...')
-
-
-            ####These functions below are private and add functionality to the UI
-
-
-    """ =============================================
-    Please do not delete the files below this line!
-    These files are the working files that allow the UI
-    to operate and do things!
-    ============================================== """
-
-    def set_module_button(self, QVBoxLayout, PushButtonName):
-        self.horizontalLayout = QtGui.QHBoxLayout()
-        self.horizontalLayout.setMargin(11)
-        self.horizontalLayout.setSpacing(6)
-        spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem)
-        PushButtonName.setObjectName(_fromUtf8("run_module_files"))
-        self.horizontalLayout.addWidget(PushButtonName)
-        QVBoxLayout.addLayout(self.horizontalLayout)
-        PushButtonName.setText(_translate("MainWindow", "Run Module", None))
-
-    def menu_item_shortcuts(self):
-        self.actionExit.setShortcut("ctrl+Q")
-
-    def menu_item_functions(self, MainWindow):
-        self.actionSet_output_location.triggered.connect(lambda: pysat_ui.output_folder(self, MainWindow))
-        self.actionLoad_Unknown_Data.triggered.connect(lambda: pysat_ui.unknown_data(self, MainWindow))
-        self.actionLoad_reference_Data.triggered.connect(lambda: pysat_ui.reference_data(self, MainWindow))
-        self.actionNormalization.triggered.connect(lambda: pysat_ui.normalization(self, MainWindow))
-        self.actionInterpolate.triggered.connect(lambda: pysat_ui.interpolated(self, MainWindow))
-        self.actionApply_Mask.triggered.connect(lambda: pysat_ui.masked(self, MainWindow))
-        self.actionStratified_Folds.triggered.connect(lambda: pysat_ui.stratified_folds(self, MainWindow))
-        self.actionTrain.triggered.connect(lambda: pysat_ui.regression_train(self, MainWindow))
-
-    def saveworkflow(self):
-        # TODO save the current window's data into a save file
+    def do_mask(self, MainWindow):
         pass
 
-    def openworkflow(self):
-        # TODO open file dialog
-        self.filename = QFileDialog.getOpenFileName(self, "Open a Workflow File", '.', "(*.wrf)")
+    def submodel_ranges(self, MainWindow):
+        pass
 
+    def do_strat_folds(self, MainWindow):
+        pass
 
-def make_combobox(choices):
-    combo = QtGui.QComboBox()
+    def do_regression_train(self, MainWindow):
+        regression_.regression_train(self, MainWindow)
 
-    for i, choice in enumerate(choices):
-        combo.addItem(_fromUtf8(""))
-        combo.setItemText(i, _translate('', choice, None))
-
-    return combo
-
-def make_listwidget(choices):
-    listwidget = QtGui.QListWidget()
-    listwidget.setItemDelegate
-    for item in choices:
-        item = QtGui.QListWidgetItem(item)
-        listwidget.addItem(item)
+    def do_scatter_plot(self, MainWindow):
+        pass
